Split tests up
