--- test/fpy/test_seqs.py
@@ -1,5755 +1,0 @@
-import pytest
-
-from fpy.types import FpyValue, U32
-
-from fpy.model import DirectiveErrorCode
-from fpy.test_helpers import (
-    assert_run_success,
-    assert_compile_failure,
-    assert_compile_success,
-    assert_run_failure,
-    lookup_type,
-)
-
-
-# When --use-gds is NOT passed (the default), override fprime_test_api with None
-# so tests run against the Python model instead of a live GDS.
-# When --use-gds IS passed, delegate to the fprime-gds plugin's session fixture
-# so tests run against the real deployment.
-@pytest.fixture(name="fprime_test_api", scope="module")
-def fprime_test_api_override(request):
-    if request.config.getoption("--use-gds"):
-        return request.getfixturevalue("fprime_test_api_session")
-    return None
-
-
-def test_comment(fprime_test_api):
-    seq = """
-# test
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_empty(fprime_test_api):
-    seq = """"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_no_newline(fprime_test_api):
-    seq = """# test"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_last_line_comment(fprime_test_api):
-    seq = """
-# test"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_simple_var(fprime_test_api):
-    seq = """
-var: U32 = 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_escaped_reserved_word(fprime_test_api):
-    # $ prefix can be used to escape reserved words like 'def', 'while', etc.
-    seq = """
-$def: U32 = 1
-$while: U32 = 2
-$if: U32 = 3
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_two_stmts_on_same_line(fprime_test_api):
-    # Two statements on the same line should fail to compile
-    seq = """
-0value: U8 = 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_no_trailing_newline(fprime_test_api):
-    # Code without a trailing newline should still compile
-    seq = "x: U32 = 1"  # No trailing newline
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_int_literal(fprime_test_api):
-    seq = """
-var: I64 = 123_456
-var = -123_456
-var = +123_456
-var = 000_00000_0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_bad_int_literal(fprime_test_api):
-    seq = """
-var: I64 = 0123_456
-
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_float_literal(fprime_test_api):
-    seq = """
-var: F32 = 1.000e-5
-var = .1
-var = 2.123
-var = 100.5e+10
-var = -123.456
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_bad_float_literal(fprime_test_api):
-    seq = """
-var: F32 = 1.
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_hex_literal(fprime_test_api):
-    seq = """
-var: U32 = 0xFF
-assert var == 255
-var = 0xDEADBEEF
-assert var == 3735928559
-var = 0x0
-assert var == 0
-var = 0X1A2B
-assert var == 6699
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_hex_literal_underscore(fprime_test_api):
-    seq = """
-var: U32 = 0xFF_FF
-assert var == 65535
-var = 0xDEAD_BEEF
-assert var == 3735928559
-var = 0x00_11_22_33
-assert var == 1122867
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_exit_success(fprime_test_api):
-    seq = """
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_exit_failure(fprime_test_api):
-    seq = """
-exit(123)
-"""
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_large_var(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_assign_to_var(fprime_test_api):
-    seq = """
-x: U32 = 1
-var: U32 = x
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nonexistent_var(fprime_test_api):
-    seq = """
-var = 1
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_namespace_type_annotation_fails(fprime_test_api):
-    seq = """
-var: Svc = 1
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_create_after_assign_var(fprime_test_api):
-    seq = """
-var = 1
-var: U32 = 2
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_bad_assign_type(fprime_test_api):
-    seq = """
-var: failure = 1
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_weird_assign_type(fprime_test_api):
-    seq = """
-var: CdhCore.cmdDisp.CMD_NO_OP = 1
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_reassign(fprime_test_api):
-    seq = """
-var: U32 = 1
-var = 2
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_reassign_ann(fprime_test_api):
-    seq = """
-var: U32 = 1
-var: U32 = 2
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_inconsistent_type(fprime_test_api):
-    seq = """
-var: U32 = 1
-var: U16 = 2
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_function_value(fprime_test_api):
-    seq = """
-var: U32 = CdhCore.cmdDisp.CMD_NO_OP
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_call_cmd(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP()
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_call_namespace_fails(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_call_cmd_with_str_arg(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP_STRING("hello world")
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_call_cmd_with_int_arg(fprime_test_api):
-    seq = """
-Ref.sendBuffComp.PARAMETER3_PRM_SET(4)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_bad_enum_ctor(fprime_test_api):
-    seq = """
-Ref.SG5.Settings(123, 0.5, 0.5, Ref.SignalType(1))
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_cmd_with_enum(fprime_test_api):
-    seq = """
-Ref.SG5.Settings(123, 0.5, 0.5, Ref.SignalType.TRIANGLE)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_instantiate_type_for_cmd(fprime_test_api):
-    seq = """
-Ref.typeDemo.CHOICE_PAIR(Ref.ChoicePair(Ref.Choice.ONE, Ref.Choice.TWO))
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_with_enum_type(fprime_test_api):
-    seq = """
-var: Ref.Choice = Ref.Choice.ONE
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_simple_if(fprime_test_api):
-    seq = """
-var: bool = True
-
-# use exit(0) if we want the sequence to succeed
-# exit(1) if we want it to fail. helpful for testing.
-
-if var:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_or_expr(fprime_test_api):
-    seq = """
-if True or False:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_not_expr(fprime_test_api):
-    seq = """
-if not False:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_or_expr_with_vars(fprime_test_api):
-    seq = """
-var1: bool = True
-var2: bool = False
-
-if var1 or var2:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_geq(fprime_test_api):
-    seq = """
-if 2 >= 1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_geq_tlm(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP()
-# NOTE! this is not guaranteed to work, if the tlm gets written
-# too slowly to the DB then this will fail
-if CdhCore.cmdDisp.CommandsDispatched >= 1:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(
-        fprime_test_api,
-        seq,
-        {"CdhCore.cmdDisp.CommandsDispatched": FpyValue(U32, 1).serialize()},
-    )
-
-
-def test_large_elifs(fprime_test_api):
-    seq = """
-if CdhCore.cmdDisp.CommandsDispatched == 0:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("0")
-elif CdhCore.cmdDisp.CommandsDispatched == 1:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("1")
-elif CdhCore.cmdDisp.CommandsDispatched == 2:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("2")
-elif CdhCore.cmdDisp.CommandsDispatched == 3:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("3")
-elif CdhCore.cmdDisp.CommandsDispatched == 4:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("4")
-else:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING(">4")
-"""
-
-    assert_run_success(
-        fprime_test_api,
-        seq,
-        {"CdhCore.cmdDisp.CommandsDispatched": FpyValue(U32, 4).serialize()},
-    )
-
-
-def test_int_as_stmt(fprime_test_api):
-    seq = """
-2
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_expr_as_stmt(fprime_test_api):
-    seq = """
-2 + 2
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_str_as_stmt(fprime_test_api):
-    seq = """
-"test"
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_complex_as_stmt(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_struct_member_of_tlm(fprime_test_api):
-    seq = """
-Ref.typeDemo.CHOICE_PAIR(Ref.ChoicePair(Ref.Choice.ONE, Ref.Choice.ONE))
-if Ref.typeDemo.ChoicePairCh.firstChoice == Ref.Choice.ONE:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(
-        fprime_test_api,
-        seq,
-        {
-            "Ref.typeDemo.ChoicePairCh": FpyValue(lookup_type(fprime_test_api, "Ref.ChoicePair"),
-                {"firstChoice": "ONE", "secondChoice": "ONE"}
-            ).serialize()
-        },
-    )
-
-
-def test_deeply_nested_loops_exhaust_recursion_depth(fprime_test_api):
-    depth = 500
-    loop_header_lines = [
-        ("    " * level) + f"for i{level} in 0 .. 1:" for level in range(depth)
-    ]
-    seq = "\n" + "\n".join(loop_header_lines) + "\n" + ("    " * depth) + "pass\n"
-
-    # Purposefully triggers RecursionError inside the compiler's parse transform.
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_const_struct_member(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-if var.priority == 3:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_member_of_anon_expr(fprime_test_api):
-    seq = """
-var: U32 = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED).priority
-if var == 3:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_time_member(fprime_test_api):
-    seq = """
-if Fw.Time(TimeBase.TB_NONE, 1, 2, 3).useconds == 3:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_cmp(fprime_test_api):
-    seq = """
-if 4.0 > 5.0:
-    exit(1)
-exit(0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_rel(fprime_test_api):
-    seq = """
-sleep(1, 1000)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_rel_default_usec(fprime_test_api):
-    seq = """
-sleep(seconds=1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_rel_default_sec(fprime_test_api):
-    seq = """
-sleep(useconds=500)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_rel_no_args(fprime_test_api):
-    seq = """
-sleep()
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_abs(fprime_test_api):
-    seq = """
-sleep_until(Fw.Time(TimeBase.TB_WORKSTATION_TIME, 0, 123, 123))
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_abs_var_arg(fprime_test_api):
-    seq = """
-x: U32 = 123
-sleep_until(Fw.Time(TimeBase.TB_WORKSTATION_TIME, 0, x, 123))
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_abs_var_arg_2(fprime_test_api):
-    seq = """
-x: Fw.Time = Fw.Time(TimeBase.TB_WORKSTATION_TIME, 1, 2, 3)
-sleep_until(x)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wait_abs_bad_arg(fprime_test_api):
-    seq = """
-sleep_until(2, 1, 2, 3)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_time_type_ctor(fprime_test_api):
-    seq = """
-var: Fw.Time = Fw.Time(TimeBase.TB_NONE, 1, 2, 3)
-if var.timeBase == TimeBase.TB_NONE and var.timeContext == 1:# and var.seconds == 2 and var.useconds == 3:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_aliases(fprime_test_api):
-    """Fw.Time is an alias for Fw.TimeValue, Fw.TimeInterval for Fw.TimeIntervalValue."""
-    seq = """
-# Fw.Time and Fw.TimeValue are interchangeable
-t1: Fw.Time = Fw.TimeValue(TimeBase.TB_NONE, 0, 100, 500000)
-t2: Fw.TimeValue = Fw.Time(TimeBase.TB_NONE, 0, 50, 0)
-assert t1.seconds == 100
-assert t2.seconds == 50
-
-# Fw.TimeInterval and Fw.TimeIntervalValue are interchangeable
-i1: Fw.TimeInterval = Fw.TimeIntervalValue(10, 500000)
-i2: Fw.TimeIntervalValue = Fw.TimeInterval(5, 0)
-assert i1.seconds == 10
-assert i2.seconds == 5
-
-# Cross-alias operations work
-result: Fw.Time = t2 + i1
-assert result.seconds == 60
-diff: Fw.TimeInterval = t1 - t2
-assert diff.seconds == 50
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_struct_ctor_var_arg(fprime_test_api):
-    seq = """
-id: U32 = 111
-priority: U32 = 3
-state: Fw.DpState = Fw.DpState.UNTRANSMITTED
-var: Svc.DpRecord = Svc.DpRecord(id, 1, 2, priority, 4, 5, state)
-if var.priority == priority and var.id == id and state == var.state:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_array_ctor_var_arg(fprime_test_api):
-    seq = """
-arr_0: U32 = 123
-arr_1: U32 = 456
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(arr_0, arr_1)
-if val[0] == arr_0 and val[1] == arr_1:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_f32_f64_cmp(fprime_test_api):
-    seq = """
-val: F32 = 0.0
-val2: F64 = 1.0
-if val > val2:
-    exit(1)
-exit(0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_construct_array(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(0, 0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_item_of_array(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(222, 111)
-if val[0] == 222:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_item_of_anon_expr(fprime_test_api):
-    seq = """
-if Svc.ComQueueDepth(123, 456)[1] == 456:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_i32_f64_cmp(fprime_test_api):
-    seq = """
-val: I32 = 2
-val2: F64 = 1.0
-if val > val2:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_i32_u32_cmp(fprime_test_api):
-    seq = """
-val: I32 = -2
-val2: U32 = 2
-# fails to compile, can't compare types of diff signedness
-if val < val2:
-    exit(1)
-exit(0)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# caught one bug
-def test_float_int_literal_cmp(fprime_test_api):
-    seq = """
-if 1 < 2.0:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assign_float_to_int(fprime_test_api):
-    seq = """
-val: I64 = 1.0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# caught one bug
-def test_and_of_ors(fprime_test_api):
-    seq = """
-if True or False and True or True:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_true(fprime_test_api):
-    seq = """
-if True:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_false(fprime_test_api):
-    seq = """
-if False:
-    exit(1)
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_else_true(fprime_test_api):
-    seq = """
-if True:
-    exit(0)
-else:
-    exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_else_false(fprime_test_api):
-    seq = """
-if False:
-    exit(1)
-else:
-    exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_elif_else(fprime_test_api):
-    seq = """
-if False:
-    exit(1)
-elif True:
-    exit(0)
-else:
-    exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_and_true_true(fprime_test_api):
-    seq = """
-if True and True:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_and_true_false(fprime_test_api):
-    seq = """
-if True and False:
-    exit(1)
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_or_false_false(fprime_test_api):
-    seq = """
-if False or False:
-    exit(1)
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_or_true_false(fprime_test_api):
-    seq = """
-if True or False:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_not_true(fprime_test_api):
-    seq = """
-if not True:
-    exit(1)
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_not_false(fprime_test_api):
-    seq = """
-if not False:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_complex_and_or_not(fprime_test_api):
-    seq = """
-if not False and (True or False):
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_literal_comparison(fprime_test_api):
-    seq = """
-if 255 > 254:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_literal_comparison_false(fprime_test_api):
-    seq = """
-if 255 < 254:
-    exit(1)
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_u8(fprime_test_api):
-    seq = """
-val1: U8 = 200
-val2: U8 = 100
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_i8(fprime_test_api):
-    seq = """
-val1: I8 = 100
-val2: I8 = -100
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_u32(fprime_test_api):
-    seq = """
-val1: U32 = 4294967295
-val2: U32 = 0
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_i32(fprime_test_api):
-    seq = """
-val1: I32 = 2147483647
-val2: I32 = -2147483648
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_f32(fprime_test_api):
-    seq = """
-val1: F32 = 3.14159
-val2: F32 = -3.14159
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_all_comparison_operators_f64(fprime_test_api):
-    seq = """
-val1: F64 = 3.14159265359
-val2: F64 = -3.14159265359
-
-if val1 > val2 and val2 < val1:
-    if val1 >= val2 and val2 <= val1:
-        if val1 != val2 and not (val1 == val2):
-            exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mixed_numeric_comparisons(fprime_test_api):
-    seq = """
-val_u8: U8 = 255
-val_i8: I8 = -10
-val_u32: U32 = 4294967295
-val_i32: I32 = -2147483648
-val_f32: F32 = 3.14159
-val_f64: F64 = -3.14159265359
-
-# i32 > u32 because the cmp happens as unsigned, and so the
-# two's complement negative is really large
-if val_u8 < val_i8 and val_i32 > val_u32:
-    if val_f64 <= val_f32 and val_f32 >= val_f64:
-        if val_u8 != val_i8 and not (val_u32 == val_i32):
-            exit(0)
-exit(1)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_equality_edge_cases(fprime_test_api):
-    seq = """
-val1: U8 = 0
-val2: U8 = 0
-val3: F32 = 0.0
-val4: F64 = 0.0
-val5: I32 = 0
-
-if val1 == val2 and val3 == val4 and val4 == val5:
-    if not (val1 != val2) and not (val3 != val4) and not (val4 != val5):
-        exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_boolean_expressions(fprime_test_api):
-    seq = """
-if not (True and False or True and not False) and True:
-    exit(1)  # Should not execute
-exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_maximum_integer_comparisons(fprime_test_api):
-    seq = """
-val_max: I64 = 9223372036854775807  # Max I64
-val_mid: I64 = 1
-val_min: I64 = -9223372036854775808  # Min I64
-
-if val_max > val_mid and val_mid > val_min:
-    if val_min < val_max:
-        exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_complex_type_assignments(fprime_test_api):
-    seq = """
-val1: I8 = 127
-val2: U8 = 255
-val3: F32 = 127.0
-
-if val1 == val3:  # Integer to float comparison
-    if val2 > val3:  # Unsigned vs float comparison
-        exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_negative_val_unsigned_type(fprime_test_api):
-    seq = """
-val1: U32 = -1
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_overflow_compile_error(fprime_test_api):
-    seq = """
-val1: U8 = 256  # Should fail: value too large for U8
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_mixed_boolean_numeric_comparison(fprime_test_api):
-    seq = """
-val1: U8 = 1
-val2: I8 = -1
-if (val1 > 0) == True and (val2 < 0) == True:  # Compare boolean results
-    if not ((val1 <= 0) == True or (val2 >= 0) == True):
-        exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_complex_boolean_nesting(fprime_test_api):
-    seq = """
-if not not not not not True:  # Multiple not operators
-    exit(1)
-elif not (True and not (False or not True)):  # Complex nesting
-    exit(1)
-else:
-    exit(0)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_non_const_str_arg(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP_STRING(Ref.cmdSeq.SeqPath)
-"""
-    # currently can't do non const string args
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_non_const_int_arg(fprime_test_api):
-    seq = """
-var: U8 = 255
-Ref.sendBuffComp.PARAMETER3_PRM_SET(var)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_non_const_float_arg(fprime_test_api):
-    seq = """
-var: F32 = 1.2
-Ref.sendBuffComp.PARAMETER4_PRM_SET(var)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_non_const_builtin_arg(fprime_test_api):
-    seq = """
-var: U32 = 1
-var2: U32 = 123123
-sleep(var, var2)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_add_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 500
-var2: U32 = 1000
-if var1 + var2 == 1500 and (var1 + 1) > var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_add_signed(fprime_test_api):
-    seq = """
-var1: I32 = -255
-var2: I32 = 255
-if var1 + var2 == 0 and (var1 + 1) > (var1 + -1):
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_add_float(fprime_test_api):
-    seq = """
-var1: F32 = -255.0
-var2: F32 = 255.0
-if var1 + var2 == 0.0 and (var1 + 1.0) > (var1 + -1.0):
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_specific_float_binop(fprime_test_api):
-    """Const folding binary ops on specific float types (F64, F32).
-
-    When both operands are cast to a specific float type (e.g. F64(1.5)),
-    their .val is a Python float, not Decimal. The const folder must handle
-    the bare `float` result from `float + float` etc.
-    """
-    seq = """
-if F64(1.5) + F64(2.5) == 4.0:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# this test inspired by a bug
-def test_float_truncate_stack_size(fprime_test_api):
-    seq = """
-var2: F64 = 123.0
-var1: F32 = F32(-var2)
-if var1 == -123.0:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_sub_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 1000
-var2: U32 = 500
-if var1 - var2 == 500 and (var1 - 1) < var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_sub_signed(fprime_test_api):
-    seq = """
-var1: I32 = 255
-var2: I32 = 255
-if var1 - var2 == 0 and (var1 - 1) < (var1 - -1):
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_sub_float(fprime_test_api):
-    seq = """
-var1: F32 = 255.0
-var2: F32 = 255.0
-if var1 - var2 == 0.0 and (var1 - 1.0) < (var1 - -1.0):
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mul_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 5
-var2: U32 = 20
-if var1 * var2 == 100 and (var1 * 2) > var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mul_signed(fprime_test_api):
-    seq = """
-var1: I32 = -5
-var2: I32 = 20
-if var1 * var2 == -100 and (var1 * 2) < var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mul_float(fprime_test_api):
-    seq = """
-var1: F32 = 5.0
-var2: F32 = 20.0
-if var1 * var2 == 100.0 and (var1 * 2.0) > var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_div_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 20
-var2: U32 = 5
-if var1 / var2 == 4.0 and (var1 / 2) < var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_div_signed(fprime_test_api):
-    seq = """
-var1: I32 = -20
-var2: I32 = 5
-if var1 / var2 == -4.0: # and (var1 / -2) > var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_div_float(fprime_test_api):
-    seq = """
-var1: F32 = -20.0
-var2: F32 = 5.0
-if var1 / var2 == -4.0 and (var1 / -2.0) > var1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# this test caught one bug (my mom spotted it)
-def test_order_of_operations(fprime_test_api):
-    seq = """
-if 1 - 2 + 3 * 4 == 11 and 10.0 / 5.0 * 2.0 == 4.0:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_arithmetic_arg_to_builtin_bad_type(fprime_test_api):
-    seq = """
-sleep(1 + 2 * 0, (0 + 1 / 2))
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_arithmetic_arg_to_builtin(fprime_test_api):
-    seq = """
-sleep(1 + 2 * 0, (0 + 1 // 2))
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_chain_mul(fprime_test_api):
-    seq = """
-var1: I32 = 1
-var2: I32 = 2
-var3: I32 = 3
-if var1 * var2 * var3 == 6:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_chain_add(fprime_test_api):
-    seq = """
-var1: I32 = 1
-var2: I32 = 2
-var3: I32 = 3
-if var1 + var2 + var3 == 6:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_chain_sub(fprime_test_api):
-    seq = """
-var1: I32 = 1
-var2: I32 = 2
-var3: I32 = 3
-if var1 - var2 - var3 == -4:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_chain_div(fprime_test_api):
-    seq = """
-var1: I32 = 3
-var2: I32 = 2
-var3: I32 = 1
-if var1 / var3 / var2 == 3/2:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_pow_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 20
-var2: U32 = 2
-if var1 ** var2 == 400:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_pow_signed(fprime_test_api):
-    seq = """
-var1: I32 = -20
-var2: I32 = 2
-if var1 ** var2 == 400:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_pow_float(fprime_test_api):
-    seq = """
-var1: F32 = 4.0
-var2: F32 = 0.5
-if var1 ** var2 == 2:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_int_literal_as_float(fprime_test_api):
-    seq = """
-var: F32 = 1
-if var == 1.0:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_log(fprime_test_api):
-    seq = """
-if log(4.0) > 1.385 and log(4.0) < 1.387:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assign_complex(fprime_test_api):
-    seq = """
-var: I64 = 1 + 1
-var = var + 3
-if var == 5:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assign_cycle(fprime_test_api):
-    seq = """
-var: I64 = var
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_cycle_2(fprime_test_api):
-    seq = """
-var: I64 = (var + 1)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_use_before_declare(fprime_test_api):
-    seq = """
-var: I64 = var2
-var2: I64 = 0
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_math_after_cmd(fprime_test_api):
-    seq = """
-var: I32 = 1
-CdhCore.cmdDisp.CMD_NO_OP()
-# making sure that the cmd doesn't mess with the stack
-if var + 1 == 2:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_cmd_return_val(fprime_test_api):
-    seq = """
-ret: Fw.CmdResponse = CdhCore.cmdDisp.CMD_NO_OP()
-if ret == Fw.CmdResponse.OK:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_struct_eq(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var2: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var3: Svc.DpRecord = Svc.DpRecord(123, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-if var == var2 and var != var3:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_complex_eq_fail(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var2: Fw.CmdResponse = Fw.CmdResponse.OK
-exit(var == var2)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_fold_struct_eq(fprime_test_api):
-    """Test that struct equality can be constant folded"""
-    seq = """
-# Both structs are constant expressions, should be folded at compile time
-if Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED) == Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED):
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_struct_neq(fprime_test_api):
-    """Test that struct inequality can be constant folded"""
-    seq = """
-# Both structs are constant expressions with different values
-if Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED) != Svc.DpRecord(123, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED):
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_enum_eq(fprime_test_api):
-    """Test that enum constant equality can be constant folded"""
-    seq = """
-# Both are enum constants, should be folded at compile time
-if Fw.DpState.UNTRANSMITTED == Fw.DpState.UNTRANSMITTED:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_enum_neq(fprime_test_api):
-    """Test that enum constant inequality can be constant folded"""
-    seq = """
-# Both are enum constants with different values
-if Fw.DpState.UNTRANSMITTED != Fw.DpState.TRANSMITTED:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_array_eq(fprime_test_api):
-    """Test that array equality can be constant folded"""
-    seq = """
-# Both arrays are constant expressions, should be folded at compile time
-if Svc.ComQueueDepth(100, 200) == Svc.ComQueueDepth(100, 200):
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_const_fold_array_neq(fprime_test_api):
-    """Test that array inequality can be constant folded"""
-    seq = """
-# Both arrays are constant expressions with different values
-if Svc.ComQueueDepth(100, 200) != Svc.ComQueueDepth(100, 300):
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mod_float(fprime_test_api):
-    seq = """
-var1: F32 = 25.25
-var2: F32 = 5
-if var1 % var2 == 0.25 and (var1 + 1) % var2 == 1.25:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mod_unsigned(fprime_test_api):
-    seq = """
-var1: U32 = 5
-var2: U32 = 20
-if var2 % var1 == 0 and (var2 + 1) % var1 == 1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_mod_signed(fprime_test_api):
-    seq = """
-var1: I32 = -5
-var2: I32 = 20
-if var2 % var1 == 0 and (var2 + 1) % var1 == -4:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-@pytest.mark.parametrize(
-    "lhs_type,rhs_type,lhs_value,rhs_value,result_type,expected_value",
-    [
-        ("U64", "I64", "9", "2", "U64", "4"),
-        ("U64", "U64", "9", "2", "U64", "4"),
-        ("F64", "F64", "5.5", "2.0", "F64", "2.0"),
-        ("F64", "F64", "-5.5", "2.0", "F64", "-2.0"),
-        ("F64", "I64", "5.5", "2", "F64", "2.0"),
-        ("I64", "F64", "5", "2.5", "F64", "2.0"),
-        ("U64", "F64", "9", "2.0", "F64", "4.0"),
-        ("F64", "U64", "9.0", "2", "F64", "4.0"),
-    ],
-)
-def test_floor_divide_64_bit_numeric_types(
-    fprime_test_api,
-    lhs_type,
-    rhs_type,
-    lhs_value,
-    rhs_value,
-    result_type,
-    expected_value,
-):
-    seq = f"""
-lhs: {lhs_type} = {lhs_value}
-rhs: {rhs_type} = {rhs_value}
-result: {result_type} = lhs // rhs
-assert result == {expected_value}
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-@pytest.mark.parametrize(
-    "lhs_type,rhs_type,lhs_value,rhs_value,result_type,expected_value",
-    [
-        ("I64", "U64", "9", "2", "U64", "4"),
-        ("U64", "I64", "9", "2", "U64", "4"),
-    ],
-)
-def test_floor_divide_64_bit_numeric_types(
-    fprime_test_api,
-    lhs_type,
-    rhs_type,
-    lhs_value,
-    rhs_value,
-    result_type,
-    expected_value,
-):
-    seq = f"""
-lhs: {lhs_type} = {lhs_value}
-rhs: {rhs_type} = {rhs_value}
-result: {result_type} = lhs // rhs
-assert result == {expected_value}
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_bool_stack_value(fprime_test_api):
-    seq = """
-if (1 == 1) == True:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_u8_too_large(fprime_test_api):
-    seq = """
-var: U8 = 123
-var = 256
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_string_eq(fprime_test_api):
-    seq = """
-exit("asdf" == "asdf")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_string_var_eq(fprime_test_api):
-    seq = """
-var: string = "test"
-var1: string = "test"
-exit(var == var1)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_string_type(fprime_test_api):
-    seq = """
-var: string = "test"
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_too_many_dirs(fprime_test_api):
-    from fpy.types import MAX_DIRECTIVES_COUNT
-
-    seq = "CdhCore.cmdDisp.CMD_NO_OP()\n" * (MAX_DIRECTIVES_COUNT + 1)
-    assert_compile_failure(fprime_test_api, seq)
-
-
-
-def test_dir_too_large(fprime_test_api):
-    # TODO this doesn't actually crash cuz the dir is too large... not sure at the moment how to trigger this
-    from fpy.types import MAX_DIRECTIVE_SIZE
-
-    seq = 'CdhCore.cmdDisp.CMD_NO_OP_STRING("' + "a" * MAX_DIRECTIVE_SIZE + '")'
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_readme_examples(fprime_test_api):
-    seq = """
-Ref.sendBuffComp.PARAMETER4_PRM_SET(1 - 2 + 3 * 4 + 10 / 5 * 2)
-param4: F32 = 15.0
-Ref.sendBuffComp.PARAMETER4_PRM_SET(param4)
-
-#prm_3: U8 = Ref.sendBuffComp.parameter3
-#cmds_dispatched: U32 = CdhCore.cmdDisp.CommandsDispatched
-cmds_dispatched: U32 = 0
-
-signal_pair: Ref.SignalPair = Ref.SignalPair(0, 0)
-
-signal_pair.time = 0.2
-
-# Svc.ComQueueDepth is an array type
-com_queue_depth: Svc.ComQueueDepth = Svc.ComQueueDepth(0, 0)
-com_queue_depth[0] = 1
-#signal_pair_time: F32 = Ref.SG1.PairOutput.time
-#com_queue_depth_0: U32 = ComCcsds.comQueue.comQueueDepth[0]
-
-int_value: U8 = 123
-float_value: F32 = int_value
-int_value = U8(float_value)
-assert int_value == 123
-
-uint: U32 = 123123
-int: I32 = I32(uint)
-assert int == 123123
-
-CdhCore.cmdDisp.CMD_NO_OP_STRING("second 0")
-# sleep for 1 second
-sleep(1)
-CdhCore.cmdDisp.CMD_NO_OP_STRING("second 1")
-# sleep for half a second
-sleep(useconds=500_000)
-
-value: bool = 1 > 2 and (3 + 4) != 5
-many_cmds_dispatched: bool = cmds_dispatched >= 123
-record1: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-record2: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-records_equal: bool = record1 == record2 # == True
-random_value: I8 = 4 # chosen by fair dice roll. guaranteed to be random
-
-if random_value < 0:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("won't happen")
-elif random_value > 0 and random_value <= 6:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("should happen!")
-else:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("uh oh...")
-
-time_interval: Fw.TimeInterval = {seconds: 15, useconds: 1000}
-
-array_var: Ref.DpDemo.U32Array = [0, 1, 2, 3, 4]
-
-counter: U64 = 0
-while counter < 100:
-    counter = counter + 1
-
-assert counter == 100
-sum: I64 = 0
-# loop i from 0 inclusive to 5 exclusive
-for i in 0 .. 5:
-    sum = sum + i
-
-assert sum == 10
-counter = 0
-while True:
-    counter = counter + 1
-    if counter == 100:
-        break
-
-assert counter == 100
-odd_numbers_sum: I64 = 0
-for i in 0 .. 10:
-    if i % 2 == 0:
-        continue
-    odd_numbers_sum = odd_numbers_sum + i
-
-assert odd_numbers_sum == 25
-
-low_bitwidth_int: U8 = 123
-high_bitwidth_int: U32 = low_bitwidth_int
-# high_bitwidth_int == 123
-low_bitwidth_float: F32 = 123.0
-high_bitwidth_float: F64 = low_bitwidth_float
-# high_bitwidth_float == 123.0
-
-# Global variable for increment example - must be declared before functions that use it
-counter_global: I64 = 0
-
-def foobar():
-    if 1 + 2 == 3:
-        CdhCore.cmdDisp.CMD_NO_OP_STRING("foo")
-
-foobar()
-
-def add_vals(a: U64, b: U64) -> U64:
-    return a + b
-    
-assert add_vals(1, 2) == 3
-
-def greet(times: I64 = 3):
-    for i in 0..times:
-        CdhCore.cmdDisp.CMD_NO_OP_STRING("hello")
-
-greet()  # uses default: prints 3 times
-greet(1) # prints once
-
-def increment():
-    counter_global = counter_global + 1
-
-increment()
-increment()
-assert counter_global == 2
-
-def recurse(limit: U64):
-    if limit == 0:
-        return
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("tick")
-    recurse(limit - 1)
-
-recurse(5) # prints "tick" 5 times
-
-
-check CdhCore.cmdDisp.CommandsDispatched > 1 persist {seconds: 1}:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("more than 30 commands for 15 seconds!")
-check CdhCore.cmdDisp.CommandsDispatched > 1 timeout now() + {seconds: 60} persist {seconds: 1}:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("more than 30 commands for 2 seconds!")
-check CdhCore.cmdDisp.CommandsDispatched > 1 timeout now() + {seconds: 60} persist {seconds: 1}:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("more than 30 commands for 2 seconds!")
-timeout:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("took more than 60 seconds :(")
-check CdhCore.cmdDisp.CommandsDispatched > 1 freq {seconds: 1}: # check every 1 second
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("more than 30 commands!")
-check CdhCore.cmdDisp.CommandsDispatched > 1
-    timeout now() + {seconds: 60}
-    persist {seconds: 1}
-    freq {seconds: 1}:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("more than 30 commands for 2 seconds!")
-timeout:
-    CdhCore.cmdDisp.CMD_NO_OP_STRING("took more than 60 seconds :(")
-
-# Time functions examples
-current_time: Fw.Time = now()
-t1: Fw.Time = now()
-sleep(seconds=1)
-t2: Fw.Time = now()
-
-assert t1 <= t2
-interval1: Fw.TimeInterval = {seconds: 5}
-interval2: Fw.TimeInterval = {seconds: 10}
-
-assert interval1 < interval2
-current: Fw.Time = {timeBase: TimeBase.TB_PROC_TIME, timeContext: 0, seconds: 100, useconds: 500000}
-offset: Fw.TimeInterval = {seconds: 60}
-assert (current + offset).seconds == 160
-start: Fw.Time = {timeBase: TimeBase.TB_PROC_TIME, timeContext: 0, seconds: 100, useconds: 0}
-end: Fw.Time = {timeBase: TimeBase.TB_PROC_TIME, timeContext: 0, seconds: 105, useconds: 500000}
-assert (end - start).seconds == 5
-"""
-    assert_run_success(
-        fprime_test_api,
-        seq,
-        {"CdhCore.cmdDisp.CommandsDispatched": FpyValue(U32, 45).serialize()},
-        timeout_s=20
-    )
-
-def test_unary_plus_unsigned(fprime_test_api):
-    seq = """
-var: U32 = 1
-if +var == var:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unary_plus_signed(fprime_test_api):
-    seq = """
-var: I32 = 1
-if +var == var:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unary_plus_float(fprime_test_api):
-    seq = """
-var: F32 = 1.0
-if +var == var:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unary_minus_signed(fprime_test_api):
-    seq = """
-var: I32 = 1
-if -var == -1:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unary_minus_float(fprime_test_api):
-    seq = """
-var: F32 = 1.0
-if -var == -1.0:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# this is an interesting case, because one side is unsigned
-# has an unsigned intermediate, so the literal is converted to
-# U64, but -1 is outside range of U64 so it fails to compile.
-# what should really happen here? TODO should -var fail to compile?
-# should intermediate type of unary minus be signed?
-def test_negative_int_literal_unsigned_op(fprime_test_api):
-    seq = """
-var: U32 = 1
-if -var == -1:
-    exit(0)
-exit(1)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_multi_arg_variable_arg_cmd(fprime_test_api):
-    seq = """
-var1: I32 = 1
-var2: F32 = 1.0
-var3: U8 = 8
-CdhCore.cmdDisp.CMD_TEST_CMD_1(var1, var2, var3)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_weird_arg_type(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP_STRING(CdhCore.cmdDisp.CMD_NO_OP_STRING)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_bad_type(fprime_test_api):
-    seq = """
-var: U32 = 1
-(var + 1)(3)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_field_with_type_ann_bad(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var.priority = 123
-if var.priority == 123:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assign_field_with_type_ann_bad_2(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var.priority: U8 = 123
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_field_before_declare(fprime_test_api):
-    seq = """
-var.priority = 123
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_array_element(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(0, 0)
-val[0] = 55
-if val[0] == 55:
-    exit(0)
-exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assign_array_element_with_type_ann_bad(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(0, 0)
-val[0]: U8 = 55
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_bad_lhs_1(fprime_test_api):
-    seq = """
-Svc.ComQueueDepth = 55
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assign_bad_lhs_2(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp.CMD_NO_OP = 55
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# TODO test deep field access/assignments (2+ levels)
-
-
-def test_assign_tlm_struct_member_bad(fprime_test_api):
-    seq = """
-Ref.cmdSeq.Debug.nextStatementOpcode = 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_set_item_of_anon_expr(fprime_test_api):
-    seq = """
-Svc.ComQueueDepth(123, 456)[1] = 456
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_set_member_of_anon_expr(fprime_test_api):
-    seq = """
-Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED).priority = 5
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_array_oob_1(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(0, 0)
-val[2] = 3
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_array_oob_2(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(123, 456)
-if val[-1] == 456:
-    exit(0)
-exit(1)
-"""
-    # TODO in the future this should work, should be the last element
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_array_oob(fprime_test_api):
-    """Out-of-bounds on a const array expression (not a variable).
-    The parent is a type constructor call, so CalculateConstExprValues
-    has a non-None parent_value. Without the bounds guard there,
-    this would crash with a Python IndexError instead of a compile error.
-    """
-    seq = """
-val: U32 = Svc.ComQueueDepth(10, 20)[2]
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_variable_array_idx(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(456, 123)
-idx: I8 = 1
-if val[idx] == 123:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_variable_array_idx_oob(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(456, 123)
-idx: I8 = 2
-if val[idx] == 123:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_get_variable_array_idx_oob_2(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(456, 123)
-idx: I8 = -1
-if val[idx] == 123:
-    exit(0)
-exit(1)
-"""
-
-    # TODO this really should also assert the failure code
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_set_variable_array_idx_oob(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(456, 123)
-idx: I8 = 2
-val[idx] = 111
-"""
-
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_set_variable_array_idx(fprime_test_api):
-    seq = """
-val: Svc.ComQueueDepth = Svc.ComQueueDepth(456, 123)
-idx: I8 = 1
-val[idx] = 111
-if val[1] == 111:
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_break_outside_loop(fprime_test_api):
-    seq = """
-break
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_continue_outside_loop(fprime_test_api):
-    seq = """
-continue
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_simple_for(fprime_test_api):
-    seq = """
-for i in 0..2:
-    pass
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_for_loop_break(fprime_test_api):
-    seq = """
-counter: I64 = 0
-for i in 0 .. 10:
-    counter = counter + 1
-    if counter == 5:
-        break
-    counter = counter + 1
-assert counter == 5
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_for_loop_continue(fprime_test_api):
-    seq = """
-counter: I64 = 0
-for i in 0 .. 10:
-    counter = counter + 1
-    continue
-    counter = counter + 1
-assert counter == 10
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_for_while_break(fprime_test_api):
-    seq = """
-counter: I64 = 0
-for i in 0 .. 10:
-    while True:
-        break
-    counter = counter + 1
-assert counter == 10
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_for_loops_break_inner(fprime_test_api):
-    seq = """
-outer_count: I64 = 0
-inner_count: I64 = 0
-for i in 0 .. 10:
-    for j in 0 .. 5:
-        inner_count = inner_count + 1
-        break
-    outer_count = outer_count + 1
-assert outer_count == 10
-assert inner_count == 10
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_for_loops_break_outer(fprime_test_api):
-    seq = """
-for i in 0 .. 10:
-    for j in 0 .. 5:
-        break
-    break
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_slightly_more_complex_for(fprime_test_api):
-    seq = """
-counter: U8 = 0
-for i in 0 .. 2:
-    if i > 2:
-        exit(1)
-    counter = U8(counter + 1)
-
-
-assert counter == 2
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_loop_var_outside_loop_after(fprime_test_api):
-    seq = """
-for i in 0 .. 7:
-    pass
-assert i == 7
-"""
-    # i is scoped to the for loop body; not visible after
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_loop_var_outside_loop_before(fprime_test_api):
-    seq = """
-i = 123
-for i in 0 .. 7:
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_loop_var_redeclare_right_type(fprime_test_api):
-    seq = """
-i: I64 = 123
-for i in 0 .. 7:
-    assert i >= 0 and i < 7
-assert i == 123
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_loop_var_redeclare_right_type_after(fprime_test_api):
-    seq = """
-for i in 0 .. 7:
-    assert i >= 0 and i < 7
-
-i: I64 = 123
-assert i == 123
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_loop_var_redeclare_in_inner_scope_func(fprime_test_api):
-    seq = """
-def test():
-    for i in 0 .. 7:
-        assert i >= 0 and i < 7
-
-i: I64 = 123
-
-assert i == 123
-
-test()
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_loop_var_redeclare_in_inner_scope_after(fprime_test_api):
-    seq = """
-def test():
-    for i in 0 .. 7:
-        pass
-
-    # After block scoping, this is fine: i is scoped to the for body
-    i: I64 = 123
-    assert i == 123
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_two_fors_same_loop_var(fprime_test_api):
-    seq = """
-for i in 0 .. 7:
-    assert i >= 0 and i < 7
-for i in 0 .. 7:
-    assert i >= 0 and i < 7
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_loop_var_redeclare_wrong_type(fprime_test_api):
-    # With block scoping, the for loop var shadows the outer i. No conflict.
-    seq = """
-i: U16 = 123
-for i in 0 .. 7:
-    pass
-assert i == 123
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_scope_override_name(fprime_test_api):
-    # With block scoping, each indentation block creates a new scope.
-    # So while/if bodies can shadow variables from outer scopes.
-    seq = """
-i: U8 = 0
-while True:
-    i: U8 = 1
-    if i == 1:
-        exit(0)
-    exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_override_global_name(fprime_test_api):
-    # Can't shadow dictionary namespaces with user variables
-    seq = """
-CdhCore: U8 = 1
-if CdhCore == 1:
-    exit(0)
-exit(1)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assert(fprime_test_api):
-    seq = """
-assert True
-assert not False
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_assert_failure(fprime_test_api):
-    seq = """
-assert False
-"""
-
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_assert_failure_with_exit_code(fprime_test_api):
-    seq = """
-assert False, 123
-"""
-
-    assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-
-def test_assert_wrong_bool_type(fprime_test_api):
-    seq = """
-assert 123
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_assert_wrong_exit_code_type(fprime_test_api):
-    seq = """
-assert True, True
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_redeclare_after_scope(fprime_test_api):
-    seq = """
-for i in 0 .. 7:
-    pass
-i: U16 = 0
-assert i == 0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_redeclare_after_for_in_if(fprime_test_api):
-    # For-loop var is scoped to the for body. After the if block, the
-    # name is free to be re-declared.
-    seq = """
-if True:
-    for i in 0 .. 7:
-        pass
-i: I64 = 0
-assert i == 0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_for_loops(fprime_test_api):
-    seq = """
-counter: U64 = 0
-z: U8 = 123
-for i in 0 .. 7:
-    for y in 20 .. 30:
-        assert i < 8
-        assert y >= 20 and y < 30
-        assert z == 123
-        counter = counter + 1
-assert counter == 70
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_redeclare_in_nested_scopes(fprime_test_api):
-    seq = """
-z: U8 = 123
-for i in 0 .. 7:
-    for z in 0 .. 7:
-        assert z < 8
-assert z == 123
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_for_loop_declare_var_bad(fprime_test_api):
-    seq = """
-for x.y in 0 .. 7:
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_use_loop_var_in_bounds(fprime_test_api):
-    seq = """
-for i in i .. 8:
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_time(fprime_test_api):
-    seq = """
-time: Fw.Time = now()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_signed_int_const_casts(fprime_test_api):
-    seq = """
-assert I8(-256) == 0
-assert I8(-129) == 127
-assert I16(-65536) == 0
-assert I16(-32769) == 32767
-assert I32(-4294967296) == 0
-assert I32(-2147483649) == 2147483647
-assert I64(-18446744073709551616) == 0
-assert I64(-9223372036854775809) == 9223372036854775807
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unsigned_int_const_casts(fprime_test_api):
-    seq = """
-assert U8(-1) == 255
-assert U8(256) == 0
-assert U16(-1) == 65535
-assert U16(65536) == 0
-assert U32(-1) == 4294967295
-assert U32(4294967296) == 0
-assert U64(-1) == 18446744073709551615
-assert U64(18446744073709551616) == 0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_const_casts(fprime_test_api):
-    seq = """
-assert F32(0.5) == 0.5
-assert F32(-0.75) == -0.75
-assert F32(1024.0) == 1024.0
-assert F32(-2048.0) == -2048.0
-assert F64(0.5) == 0.5
-assert F64(-0.75) == -0.75
-assert F64(123456789.5) == 123456789.5
-assert F64(-987654321.25) == -987654321.25
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_to_signed_int_const_casts(fprime_test_api):
-    seq = """
-assert I8(-128.0) == -128
-assert I8(127.0) == 127
-assert I16(-32768.0) == -32768
-assert I16(32767.0) == 32767
-assert I32(-2147483648.0) == -2147483648
-assert I32(2147483647.0) == 2147483647
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_to_unsigned_int_const_casts(fprime_test_api):
-    seq = """
-assert U8(0.0) == 0
-assert U8(255.0) == 255
-assert U16(65535.0) == 65535
-assert U32(4294967295.0) == 4294967295
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_signed_int_to_float_const_casts(fprime_test_api):
-    seq = """
-assert F32(-128) == -128.0
-assert F64(-128) == -128.0
-assert F32(32767) == 32767.0
-assert F64(32767) == 32767.0
-assert F64(-2147483648) == -2147483648.0
-assert F64(2147483647) == 2147483647.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unsigned_int_to_float_const_casts(fprime_test_api):
-    seq = """
-assert F32(U32(0)) == 0.0
-assert F64(U32(0)) == 0.0
-assert F32(U32(65535)) == 65535.0
-assert F64(U32(65535)) == 65535.0
-assert F64(U64(4294967295)) == 4294967295.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_signed_int_runtime_casts(fprime_test_api):
-    seq = """
-src: I64 = -256
-assert I8(src) == 0
-
-src = -129
-assert I8(src) == 127
-
-src = -65536
-assert I16(src) == 0
-
-src = -32769
-assert I16(src) == 32767
-
-src = -4294967296
-assert I32(src) == 0
-
-src = -2147483649
-assert I32(src) == 2147483647
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unsigned_int_runtime_casts(fprime_test_api):
-    seq = """
-signed_src: I64 = -1
-assert U8(signed_src) == 255
-assert U16(signed_src) == 65535
-assert U32(signed_src) == 4294967295
-assert U64(signed_src) == 18446744073709551615
-
-unsigned_src: U64 = 256
-assert U8(unsigned_src) == 0
-
-unsigned_src = 65536
-assert U16(unsigned_src) == 0
-
-unsigned_src = 4294967296
-assert U32(unsigned_src) == 0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_runtime_casts(fprime_test_api):
-    seq = """
-wide_src: F64 = 0.5
-assert F32(wide_src) == 0.5
-
-wide_src = -0.75
-assert F32(wide_src) == -0.75
-
-wide_src = 1024.0
-assert F32(wide_src) == 1024.0
-
-wide_src = -2048.0
-assert F32(wide_src) == -2048.0
-
-narrow_src: F32 = 0.5
-assert F64(narrow_src) == 0.5
-
-narrow_src = -0.75
-assert F64(narrow_src) == -0.75
-
-narrow_src = 123.5
-assert F64(narrow_src) == 123.5
-
-narrow_src = -987.25
-assert F64(narrow_src) == -987.25
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_to_signed_int_runtime_casts(fprime_test_api):
-    seq = """
-f_src: F64 = -128.0
-assert I8(f_src) == -128
-
-f_src = 127.0
-assert I8(f_src) == 127
-
-f_src = -32768.0
-assert I16(f_src) == -32768
-
-f_src = 32767.0
-assert I16(f_src) == 32767
-
-f_src = -2147483648.0
-assert I32(f_src) == -2147483648
-
-f_src = 2147483647.0
-assert I32(f_src) == 2147483647
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_float_to_unsigned_int_runtime_casts(fprime_test_api):
-    seq = """
-f_src: F64 = 0.0
-assert U8(f_src) == 0
-
-f_src = 255.0
-assert U8(f_src) == 255
-
-f_src = 65535.0
-assert U16(f_src) == 65535
-
-f_src = 4294967295.0
-assert U32(f_src) == 4294967295
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_signed_int_to_float_runtime_casts(fprime_test_api):
-    seq = """
-i_src: I32 = -128
-assert F32(i_src) == -128.0
-assert F64(i_src) == -128.0
-
-i_src = 32767
-assert F32(i_src) == 32767.0
-assert F64(i_src) == 32767.0
-
-wide_src: I64 = -2147483648
-assert F64(wide_src) == -2147483648.0
-
-wide_src = 2147483647
-assert F64(wide_src) == 2147483647.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_unsigned_int_to_float_runtime_casts(fprime_test_api):
-    seq = """
-u_src: U32 = 0
-assert F32(u_src) == 0.0
-assert F64(u_src) == 0.0
-
-u_src = 65535
-assert F32(u_src) == 65535.0
-assert F64(u_src) == 65535.0
-
-wide_src: U64 = 4294967295
-assert F64(wide_src) == 4294967295.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_downcast(fprime_test_api):
-    seq = """
-i: U32 = 123123
-u: U8 = U8(i)
-assert u == (i % 256)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_downcast_fail(fprime_test_api):
-    seq = """
-i: U32 = 123123
-u: U8 = i
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_upcast(fprime_test_api):
-    seq = """
-i: U8 = 255
-u: U32 = U32(i)
-assert u == i
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wrong_bool_type(fprime_test_api):
-    seq = """
-val: bool = 123
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_while_break_in_if(fprime_test_api):
-    seq = """
-while True:
-    if True:
-        break
-    exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_for_break_in_if(fprime_test_api):
-    seq = """
-for i in 0 .. 100:
-    if True:
-        break
-    exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_while_continue_in_if(fprime_test_api):
-    """Test that continue both skips rest of body AND loops back to condition."""
-    seq = """
-i: U64 = 0
-while i < 5:
-    i = i + 1
-    if True:
-        continue
-    exit(1)  # should be skipped by continue
-
-# Verify loop ran all 5 iterations (not just 1 like a break would)
-assert i == 5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_for_continue_in_if(fprime_test_api):
-    """Test that continue both skips rest of body AND continues to next iteration."""
-    seq = """
-sum: U64 = 0
-for i in 0 .. 100:
-    sum = sum + 1
-    if True:
-        continue
-    exit(1)  # should be skipped by continue
-
-# Verify loop ran all 100 iterations
-assert sum == 100
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_downcast_large_literal(fprime_test_api):
-    seq = """
-val: U8 = U8(1231231231243) # this is allowed but suspicious
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-@pytest.mark.xfail(reason="Support for non utf-8 characters should be added later")
-def test_non_utf_8(fprime_test_api):
-    seq = """
-val: F64 = 0.0 
-
-CdhCore.cmdDisp.CMD_NO_OP_STRING("в")
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# TODO assert failure should split based on whether it's a syntax or semantic failure
-
-
-def test_abs_float(fprime_test_api):
-    seq = """
-assert fabs(1.0) == 1.0
-assert fabs(-1.0) == 1.0
-assert fabs(0.0) == 0.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_abs_i64(fprime_test_api):
-    seq = """
-assert iabs(I64(-1)) == 1
-assert iabs(I64(1)) == 1
-assert iabs(I64(0)) == 0
-# need to use a large subtract here cuz otherwise float precision kills us... this is kinda sus
-assert iabs(I64(2**63 - 6556)) == 2**63 - 6556
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_abs_u64(fprime_test_api):
-    seq = """
-# fails, iabs takes signed
-assert iabs(U64(1)) == 1
-assert iabs(U64(0)) == 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_abs_literal_int(fprime_test_api):
-    seq = """
-assert iabs(1) == 1
-assert iabs(-1) == 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_abs_literal_float(fprime_test_api):
-    seq = """
-assert fabs(1.0) == 1.0
-assert fabs(-1.0) == 1.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_type_ann_bad(fprime_test_api):
-    seq = """
-var: Fw.Time.asdf = 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_var_type_ann_bad_2(fprime_test_api):
-    seq = """
-var: Svc = 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_folding_time_eq(fprime_test_api):
-    seq = """
-assert Fw.Time(TimeBase.TB_NONE, 0, 0, 0) == Fw.Time(TimeBase.TB_NONE, 0, 0, 0)
-assert Fw.Time(TimeBase.TB_NONE, 0, 1, 0) != Fw.Time(TimeBase.TB_NONE, 0, 0, 0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_calling_struct_field_should_fail_gracefully(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-var.priority()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_calling_namespace_should_fail_gracefully(fprime_test_api):
-    seq = """
-Ref.typeDemo()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_calling_variable_should_fail_gracefully(fprime_test_api):
-    seq = """
-x: U32 = 1
-x()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_divide_by_zero(fprime_test_api):
-    seq = """
-1 / 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_complex_pow(fprime_test_api):
-    seq = """
-(-1) ** 0.5
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_very_large_const_pow(fprime_test_api):
-    seq = """
-10.0 ** 1000
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_func_call_a_namespace(fprime_test_api):
-    seq = """
-CdhCore.cmdDisp()
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_missing_struct_member(fprime_test_api):
-    seq = """
-record: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-value: U32 = record.missing_field
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_item_of_struct(fprime_test_api):
-    seq = """
-record: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-value: U32 = record[0]
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_item_of_namespace(fprime_test_api):
-    seq = """
-value: U32 = CdhCore.cmdDisp[0]
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_literal_float_coercion_overflow(fprime_test_api):
-    seq = """
-var: F32 = 999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_empty_range(fprime_test_api):
-    seq = """
-for i in 7..0:
-    exit(1)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_simple_func_def(fprime_test_api):
-    seq = """
-def test():
-    pass
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_def_with_args(fprime_test_api):
-    seq = """
-def test(arg: U8):
-    pass
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_name_special_chars(fprime_test_api):
-    # Variable names with invalid special characters should fail
-    seq = """
-@invalid: U8 = 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_newline_in_body(fprime_test_api):
-    seq = """
-if True:
-    val: U8 = 0
-
-    pass
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_return_outside_func(fprime_test_api):
-    seq = """
-return
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_simple_return(fprime_test_api):
-    seq = """
-def test():
-    return
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_return_val(fprime_test_api):
-    seq = """
-def test() -> U8:
-    return 1
-
-assert test() == 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_wrong_return_type(fprime_test_api):
-    seq = """
-def test() -> U8:
-    return 1.0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_wrong_arg_type(fprime_test_api):
-    seq = """
-def test(arg: U8):
-    pass
-
-test(1.0)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_break_in_func_in_loop(fprime_test_api):
-    seq = """
-for i in 0..2:
-    def test(arg: U8):
-        break
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_outside_var_in_func(fprime_test_api):
-    seq = """
-i: U8 = 1
-def test():
-    assert i == 1
-test()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_outside_struct_member_in_func(fprime_test_api):
-    seq = """
-var: Svc.DpRecord = Svc.DpRecord(0, 1, 2, 3, 4, 5, Fw.DpState.UNTRANSMITTED)
-def test():
-    assert var.priority == 3
-test()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_outside_array_element_const_index_in_func(fprime_test_api):
-    """Test that functions can access array elements of global variables with a constant index"""
-    seq = """
-arr: Svc.ComQueueDepth = Svc.ComQueueDepth(123, 456)
-def test():
-    assert arr[0] == 123
-    assert arr[1] == 456
-test()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_outside_array_element_non_const_index_in_func(fprime_test_api):
-    """Test that functions can access array elements of global variables with a non-constant index"""
-    seq = """
-arr: Svc.ComQueueDepth = Svc.ComQueueDepth(123, 456)
-def test():
-    idx: I64 = 1
-    assert arr[idx] == 456
-test()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_modify_global_var_in_func(fprime_test_api):
-    """Test that functions can modify top-level (global) variables"""
-    seq = """
-i: I64 = 0
-def increment():
-    i = i + 1
-increment()
-increment()
-assert i == 2
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_modify_global_var_in_func_before_definition(fprime_test_api):
-    """Test that functions can modify top-level (global) variables, declared after the function definition"""
-    seq = """
-increment()
-
-def increment():
-    assert i == 0
-    i = i + 1
-
-i: I64 = 123
-
-assert i == 123
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_use_lvar_from_func_outside_func(fprime_test_api):
-    seq = """
-def test():
-    i: U8 = 0
-assert i == 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_use_arg_from_func_outside_func(fprime_test_api):
-    seq = """
-def test(arg: U8):
-    pass
-if arg == 0:
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_in_func(fprime_test_api):
-    seq = """
-def test(arg: U8):
-    def test2() -> U8:
-        return 0
-    assert test2() == 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_in_if(fprime_test_api):
-    """Function definitions are only allowed at the top level, not inside if blocks"""
-    seq = """
-if True:
-    def test():
-        pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_in_for(fprime_test_api):
-    """Function definitions are only allowed at the top level, not inside for loops"""
-    seq = """
-for i in 0..10:
-    def test():
-        pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_in_while(fprime_test_api):
-    """Function definitions are only allowed at the top level, not inside while loops"""
-    seq = """
-while True:
-    def test():
-        pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_use_func_outside_scope(fprime_test_api):
-    # This test previously tested nested function scoping, but nested functions
-    # are now disallowed at the syntax level. Keep as a syntax error test.
-    seq = """
-def test(arg: U8):
-    def test2() -> U8:
-        return 0
-
-assert test2() == 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_call_func(fprime_test_api):
-    seq = """
-def test() -> U8:
-    return 1
-
-def test2() -> U8:
-    return test()
-
-assert test2() == 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_func_call_func_before_defined(fprime_test_api):
-    seq = """
-
-def test2() -> U8:
-    return test()
-
-def test() -> U8:
-    return 1
-
-assert test2() == 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_two_func_args_same_name(fprime_test_api):
-    seq = """
-
-def test(arg: U8, arg: U8):
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_redeclare_func(fprime_test_api):
-    seq = """
-
-def test():
-    pass
-def test():
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_redeclare_func_from_var(fprime_test_api):
-    # Functions and variables are in separate scopes, so this is allowed
-    seq = """
-
-test: U8 = 0
-def test():
-    pass
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_redeclare_var_from_func(fprime_test_api):
-    # Functions and variables are in separate scopes, so this is allowed
-    seq = """
-
-def test():
-    pass
-test: U8 = 0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_fib(fprime_test_api):
-    seq = """
-def fib(a: U64) -> U64:
-    if a < 2:
-        return 1
-    return fib(a - 1) + fib(a - 2)
-
-assert fib(0) == 1
-assert fib(1) == 1
-assert fib(2) == 2
-assert fib(4) == 5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_var_in_func(fprime_test_api):
-    seq = """
-def test():
-    var: U32 = 123
-    assert var == 123
-
-test()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_missing_return(fprime_test_api):
-    seq = """
-def test() -> U32:
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_if_elif_return(fprime_test_api):
-    seq = """
-def test() -> U32:
-    if True:
-        return 1
-    elif False:
-        return 3
-    else:
-        return 2
-    
-
-
-assert test() == 1
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_if_elif_return_missing(fprime_test_api):
-    seq = """
-def test() -> U32:
-    if True:
-        return 1
-    elif False:
-        pass
-    else:
-        return 2
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_return_in_for(fprime_test_api):
-    seq = """
-def test() -> U32:
-    for i in 0..0: # this fails because we can't guarantee that the loop body executes
-        return 0
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_if_without_else_missing_return(fprime_test_api):
-    seq = """
-def missing_return(flag: bool) -> U32:
-    if flag:
-        return 1
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_func_with_multiple_args(fprime_test_api):
-    seq = """
-def add_vals(lhs: U32, rhs: U32) -> U64:
-    return lhs + rhs
-
-assert add_vals(1, 2) == 3
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_nested_func_capture_outer_local(fprime_test_api):
-    seq = """
-def outer() -> U32:
-    local_val: U32 = 42
-    def inner() -> U32:
-        return local_val
-    return inner()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_nested_func_capture_outer_arg(fprime_test_api):
-    seq = """
-def outer(value: U32) -> U32:
-    def inner() -> U32:
-        return value
-    return inner()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_const_float_downcast_rounding(fprime_test_api):
-    rounding_seq = """
-narrow: F32 = 16777217.0
-assert F64(narrow) == 16777216.0
-"""
-
-    assert_run_success(fprime_test_api, rounding_seq)
-
-
-def test_const_float_overflow(fprime_test_api):
-    overflow_seq = """
-narrow: F32 = 3.5e38
-"""
-
-    assert_compile_failure(fprime_test_api, overflow_seq)
-
-
-def test_and_short_circuit_skips_rhs(fprime_test_api):
-    seq = """
-def boom() -> bool:
-    assert False
-    return True
-
-if False and boom():
-    exit(1)
-exit(0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_or_short_circuit_skips_rhs(fprime_test_api):
-    seq = """
-def boom() -> bool:
-    assert False
-    return False
-
-if True or boom():
-    exit(0)
-exit(1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_void_function_without_explicit_return(fprime_test_api):
-    seq = """
-def noop():
-    pass
-
-noop()
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_param_name_matching_type(fprime_test_api):
-    seq = """
-def echo(U32: U32) -> U32:
-    return U32
-
-assert echo(5) == 5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_simple(fprime_test_api):
-    seq = """
-def test(a: U64, b: U64 = 5) -> U64:
-    return a + b
-
-assert test(1) == 6
-assert test(1, 2) == 3
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_before_non_default(fprime_test_api):
-    seq = """
-def test(a: U64 = 5, b: U64):
-    pass
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_multiple(fprime_test_api):
-    seq = """
-def test(a: U64, b: U64 = 2, c: U64 = 3) -> U64:
-    return a + b + c
-
-assert test(1) == 6
-assert test(1, 5) == 9
-assert test(1, 5, 10) == 16
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_too_few_args(fprime_test_api):
-    seq = """
-def test(a: U64, b: U64 = 5) -> U64:
-    return a + b
-
-test()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_too_many_args(fprime_test_api):
-    seq = """
-def test(a: U64, b: U64 = 5) -> U64:
-    return a + b
-
-test(1, 2, 3)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_all_defaults(fprime_test_api):
-    seq = """
-def test(a: U64 = 1, b: U64 = 2) -> U64:
-    return a + b
-
-assert test() == 3
-assert test(10) == 12
-assert test(10, 20) == 30
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_float_coercion(fprime_test_api):
-    """Float literal coercion for default value."""
-    seq = """
-def test(x: F64 = 2.5) -> F64:
-    return x + 1.0
-
-# 2.5 is an arbitrary precision Float literal that gets coerced to F64
-assert test() == 3.5
-assert test(10.0) == 11.0
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_expression(fprime_test_api):
-    """Default value can be an expression."""
-    seq = """
-def test(a: U64 = 2 + 3) -> U64:
-    return a
-
-assert test() == 5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_incompatible_type(fprime_test_api):
-    """Default value type must be compatible with parameter type."""
-    seq = """
-def test(a: U64 = -5) -> U64:
-    return a
-"""
-
-    # Should fail because -5 can't be assigned to U64
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_wrong_type(fprime_test_api):
-    """Default value type must be assignable to parameter type."""
-    seq = """
-def test(a: bool = 5) -> bool:
-    return a
-"""
-
-    # Should fail because int literal is not a bool
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_with_cast(fprime_test_api):
-    """Default value can be a cast expression."""
-    seq = """
-def test(a: U8 = U8(255)) -> U8:
-    return a
-
-assert test() == 255
-assert test(U8(10)) == 10
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_with_narrowing_cast(fprime_test_api):
-    """Default value can be a narrowing cast expression."""
-    seq = """
-def test(a: U8 = U8(1000)) -> U8:
-    return a
-
-# 1000 truncated to U8 is 232 (1000 & 0xFF)
-assert test() == 232
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_with_tlm(fprime_test_api):
-    """Default value cannot be a runtime value like telemetry - must be const expr."""
-    seq = """
-def test(a: U32 = CdhCore.cmdDisp.CommandsDispatched) -> U32:
-    return a
-"""
-
-    # Should fail because telemetry is not a const expression
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_with_function_call(fprime_test_api):
-    """Default value cannot be a function call - must be const expr."""
-    seq = """
-def helper() -> U64:
-    return 42
-
-def test(a: U64 = helper()) -> U64:
-    return a
-"""
-
-    # Should fail because function call is not a const expression
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_forward_called_function(fprime_test_api):
-    """Default arg const values are calculated even for forward-called functions.
-
-    This tests that when a function is called before it's defined in the source,
-    the default argument's const value is still properly available.
-    """
-    seq = """
-def caller() -> U64:
-    # Call test() before it's defined, using default arg
-    return test()
-
-def test(a: U64 = 42) -> U64:
-    return a
-
-assert caller() == 42
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_default_arg_with_variable(fprime_test_api):
-    """Default value cannot reference a variable - must be const expr."""
-    seq = """
-x: U64 = 10
-
-def test(a: U64 = x) -> U64:
-    return test()
-"""
-
-    # Should fail because variable is not a const expression
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_nested_func_cannot_access_outer_local(fprime_test_api):
-    """Default value cannot reference a variable - must be const expr."""
-    seq = """
-def outer() -> U64:
-    x: U64 = 10
-    
-    def inner(a: U64 = x) -> U64:
-        return a
-    
-    return inner()
-"""
-
-    # Should fail because x is not a const expression
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_forward_reference(fprime_test_api):
-    """Default value cannot reference a variable declared after the function."""
-    seq = """
-def test(a: U64 = x) -> U64:
-    return a
-
-x: U64 = 10
-"""
-
-    # Should fail: "'x' used before declared"
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_default_arg_undefined_variable(fprime_test_api):
-    """Default value cannot reference an undefined variable."""
-    seq = """
-def test(a: U64 = undefined_var) -> U64:
-    return a
-"""
-
-    # Should fail: "Unknown value"
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# Named arguments tests
-
-
-def test_named_arg_simple(fprime_test_api):
-    """Basic named argument usage."""
-    seq = """
-def test(a: U64, b: U64) -> U64:
-    return a + b
-
-assert test(a=1, b=2) == 3
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_reorder(fprime_test_api):
-    """Named arguments can be in any order."""
-    seq = """
-def test(a: U64, b: U64, c: U64) -> U64:
-    return a * 100 + b * 10 + c
-
-assert test(c=3, a=1, b=2) == 123
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_mixed_positional(fprime_test_api):
-    """Positional args followed by named args."""
-    seq = """
-def test(a: U64, b: U64, c: U64) -> U64:
-    return a * 100 + b * 10 + c
-
-assert test(1, c=3, b=2) == 123
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_with_defaults(fprime_test_api):
-    """Named arguments with default values."""
-    seq = """
-def test(a: U64, b: U64 = 5, c: U64 = 10) -> U64:
-    return a + b + c
-
-# Only provide first and last, middle uses default
-assert test(a=1, c=20) == 26
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_all_with_defaults(fprime_test_api):
-    """Named arguments for all parameters, some with defaults."""
-    seq = """
-def test(a: U64 = 1, b: U64 = 2, c: U64 = 3) -> U64:
-    return a * 100 + b * 10 + c
-
-# Override middle one only
-assert test(b=5) == 153
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_unknown_name(fprime_test_api):
-    """Error when using unknown argument name."""
-    seq = """
-def test(a: U64, b: U64) -> U64:
-    return a + b
-
-test(a=1, c=2)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_named_arg_duplicate(fprime_test_api):
-    """Error when same argument specified twice."""
-    seq = """
-def test(a: U64, b: U64) -> U64:
-    return a + b
-
-test(a=1, a=2)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_named_arg_positional_and_named(fprime_test_api):
-    """Error when same argument specified by position and by name."""
-    seq = """
-def test(a: U64, b: U64) -> U64:
-    return a + b
-
-test(1, a=2)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_named_arg_positional_after_named(fprime_test_api):
-    """Error when positional argument follows named argument."""
-    seq = """
-def test(a: U64, b: U64, c: U64) -> U64:
-    return a + b + c
-
-test(a=1, 2, 3)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_named_arg_missing_required(fprime_test_api):
-    """Error when required argument is missing."""
-    seq = """
-def test(a: U64, b: U64, c: U64) -> U64:
-    return a + b + c
-
-test(a=1, c=3)
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_named_arg_cmd_call(fprime_test_api):
-    """Named arguments work with command calls."""
-    seq = """
-CdhCore.cmdDisp.CMD_TEST_CMD_1(arg1=1, arg2=1.0, arg3=1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_cmd_call_reorder(fprime_test_api):
-    """Named arguments can reorder command arguments."""
-    seq = """
-CdhCore.cmdDisp.CMD_TEST_CMD_1(arg3=1, arg1=1, arg2=1.0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_type_ctor(fprime_test_api):
-    """Named arguments work with type constructors."""
-    seq = """
-time: Fw.Time = Fw.Time(seconds=123, useconds=456, timeBase=TimeBase.TB_NONE, timeContext=0)
-assert time.seconds == 123
-assert time.useconds == 456
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_struct_ctor(fprime_test_api):
-    """Named arguments work with struct constructors."""
-    seq = """
-pair: Ref.ChoicePair = Ref.ChoicePair(secondChoice=Ref.Choice.TWO, firstChoice=Ref.Choice.ONE)
-assert pair.firstChoice == Ref.Choice.ONE
-assert pair.secondChoice == Ref.Choice.TWO
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_builtin(fprime_test_api):
-    """Named arguments work with builtin functions."""
-    seq = """
-sleep(useconds=1000, seconds=1)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_builtin_single(fprime_test_api):
-    """Named arguments work with single-arg builtins."""
-    seq = """
-exit(exit_code=0)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_cast(fprime_test_api):
-    """Named arguments work with type casts."""
-    seq = """
-val: F64 = 3.14
-result: U32 = U32(value=val)
-assert result == 3
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_coercion_int_to_float(fprime_test_api):
-    """Named arguments correctly coerce int to float."""
-    seq = """
-def test(x: F64) -> F64:
-    return x + 0.5
-
-# U32 should be coerced to F64
-val: U32 = 10
-assert test(x=val) == 10.5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_coercion_narrow_to_wide(fprime_test_api):
-    """Named arguments correctly coerce narrow int to wider int."""
-    seq = """
-def test(x: U64) -> U64:
-    return x + 1
-
-small: U8 = 100
-# U8 should be coerced to U64
-assert test(x=small) == 101
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_coercion_in_cmd(fprime_test_api):
-    """Named arguments with coercion in command calls."""
-    seq = """
-# arg1 is I32, arg2 is F32, arg3 is U8
-# Test coercion from narrower to wider types
-val1: I16 = 1  # I16 -> I32
-val2: I16 = 2  # I16 -> F32
-val3: U8 = 3   # U8 -> U8 (exact match)
-CdhCore.cmdDisp.CMD_TEST_CMD_1(arg1=val1, arg2=val2, arg3=val3)
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_named_arg_coercion_reordered(fprime_test_api):
-    """Named arguments maintain correct coercion when reordered."""
-    seq = """
-def test(a: U64, b: F64, c: I32) -> F64:
-    return b + F64(a) + F64(c)
-
-# Pass in different order with finite bitwidth types
-x: I8 = -5
-y: U32 = 10
-z: F32 = 0.5
-assert test(c=x, a=y, b=z) == 5.5
-"""
-
-    assert_run_success(fprime_test_api, seq)
-
-
-# ==================== Check Statement Tests ====================
-
-
-def test_check_condition_true_immediately(fprime_test_api):
-    """Test that check succeeds immediately when condition is true with zero persist."""
-    seq = """
-check_passed: bool = False
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 100000):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_timeout_when_always_false(fprime_test_api):
-    """Test that check times out when condition is always false."""
-    seq = """
-timed_out: bool = False
-check False timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    assert False, 1
-timeout:
-    timed_out = True
-assert timed_out
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_condition_evaluated_multiple_times(fprime_test_api):
-    """Test that check evaluates condition multiple times until it becomes true."""
-    seq = """
-# Count how many times condition is evaluated
-eval_count: I64 = 0
-
-def check_condition() -> bool:
-    eval_count = eval_count + 1
-    # Return true on 3rd evaluation
-    return eval_count >= 3
-
-check check_condition() timeout time_add(now(), Fw.TimeIntervalValue(5, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-timeout:
-    assert False, 1
-
-# Should have been evaluated at least 3 times
-assert eval_count >= 3
-"""
-    assert_run_success(fprime_test_api, seq, timeout_s=6)
-
-
-def test_check_condition_must_persist(fprime_test_api):
-    """Test that condition must remain true for the full persist duration.
-
-    If condition becomes false before persist duration, the timer resets.
-    """
-    seq = """
-# Condition returns true twice, then false, then true forever
-call_count: I64 = 0
-
-def flaky_condition() -> bool:
-    call_count = call_count + 1
-    # True for calls 1-2, false for call 3, true thereafter
-    if call_count <= 2:
-        return True
-    if call_count == 3:
-        return False
-    return True
-
-# Calls 1-2 are true but call 3 is false, resetting the persist timer
-# Call 4+ are true, so it should eventually succeed
-check flaky_condition() timeout time_add(now(), Fw.TimeIntervalValue(10, 0)) persist Fw.TimeIntervalValue(3, 0) freq Fw.TimeIntervalValue(1, 0):
-    pass
-timeout:
-    assert False, 1
-
-# Should have been called more than 3 times since persist timer was reset
-assert call_count > 3
-"""
-    assert_run_success(fprime_test_api, seq, timeout_s=20)
-
-
-def test_check_zero_persist_true_once_enough(fprime_test_api):
-    """Test that with zero persist, condition being true once is enough."""
-    seq = """
-# Return true only once, then false forever
-returned_true: bool = False
-
-def return_true_once() -> bool:
-    if not returned_true:
-        returned_true = True
-        return True
-    return False
-
-check return_true_once() timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    exit(0)
-timeout:
-    assert False, 1
-"""
-    assert_run_success(fprime_test_api, seq, timeout_s=6)
-
-
-def test_check_body_runs_on_success(fprime_test_api):
-    """Test that check body runs when condition succeeds."""
-    seq = """
-body_ran: bool = False
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    body_ran = True
-assert body_ran
-"""
-    assert_run_success(fprime_test_api, seq, timeout_s=6)
-
-
-def test_check_timeout_body_runs_on_timeout(fprime_test_api):
-    """Test that timeout body runs when check times out."""
-    seq = """
-timeout_body_ran: bool = False
-check False timeout time_add(now(), Fw.TimeIntervalValue(0, 50000)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    assert False, 1
-timeout:
-    timeout_body_ran = True
-assert timeout_body_ran
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_optional_timeout_body(fprime_test_api):
-    """Test that check works without timeout body."""
-    seq = """
-result: bool = False
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 100000):
-    result = True
-assert result
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_no_timeout_clause(fprime_test_api):
-    """Test check without timeout clause (runs indefinitely until success)."""
-    seq = """
-# Without timeout, check runs until condition succeeds
-call_count: I64 = 0
-
-def eventually_true() -> bool:
-    call_count = call_count + 1
-    return call_count >= 3
-
-check eventually_true():
-    exit(0)
-"""
-    assert_run_success(fprime_test_api, seq, timeout_s=10)
-
-
-def test_check_only_timeout_specified(fprime_test_api):
-    """Test check with only timeout specified (uses default persist=0 and freq=1s)."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    exit(0)
-timeout:
-    assert False, 1
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_nested_in_function(fprime_test_api):
-    """Test that check statements work inside functions."""
-    seq = """
-def do_check() -> bool:
-    result: bool = False
-    check True timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-        result = True
-    timeout:
-        pass
-    return result
-
-assert do_check()
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_modifies_outer_scope(fprime_test_api):
-    """Test that check body can modify variables in outer scope."""
-    seq = """
-outer_var: I32 = 0
-
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    outer_var = 42
-
-assert outer_var == 42
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_absolute_timeout(fprime_test_api):
-    """Test that check works with absolute Fw.Time timeout."""
-    seq = """
-# Create an absolute timeout 100ms from now
-abs_timeout: Fw.Time = time_add(now(), Fw.TimeIntervalValue(0, 100000))
-result: bool = False
-check True timeout abs_timeout persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    result = True
-timeout:
-    pass
-assert result
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ==================== Check Statement Type Error Tests ====================
-
-
-def test_check_timeout_wrong_type(fprime_test_api):
-    """Test that check timeout with wrong type gives compile error."""
-    seq = """
-check True timeout 123 persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_condition_wrong_type(fprime_test_api):
-    """Test that check condition must be bool, not int."""
-    seq = """
-check 123 timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_condition_wrong_type_string(fprime_test_api):
-    """Test that check condition must be bool, not string."""
-    seq = """
-check "hello" timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_persist_wrong_type(fprime_test_api):
-    """Test that check persist must be TimeIntervalValue, not int."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist 123 freq Fw.TimeIntervalValue(0, 10000):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_persist_wrong_type_time(fprime_test_api):
-    """Test that check persist must be TimeIntervalValue, not Fw.Time."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist now() freq Fw.TimeIntervalValue(0, 10000):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_freq_wrong_type(fprime_test_api):
-    """Test that check freq must be TimeIntervalValue, not int."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq 123:
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_freq_wrong_type_time(fprime_test_api):
-    """Test that check freq must be TimeIntervalValue, not Fw.Time."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq now():
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_duplicate_timeout(fprime_test_api):
-    """Test that duplicate timeout clauses cause a compile error."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) timeout time_add(now(), Fw.TimeIntervalValue(2, 0)):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_duplicate_persist(fprime_test_api):
-    """Test that duplicate persist clauses cause a compile error."""
-    seq = """
-check True persist Fw.TimeIntervalValue(1, 0) persist Fw.TimeIntervalValue(2, 0):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_duplicate_freq(fprime_test_api):
-    """Test that duplicate freq clauses cause a compile error."""
-    seq = """
-check True freq Fw.TimeIntervalValue(1, 0) freq Fw.TimeIntervalValue(2, 0):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# ==================== Check Statement Multi-line Syntax Tests ====================
-
-
-def test_check_multiline_timeout_only(fprime_test_api):
-    """Test multi-line check with only timeout clause."""
-    seq = """
-check_passed: bool = False
-check True
-    timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_multiline_all_clauses(fprime_test_api):
-    """Test multi-line check with all three clauses."""
-    seq = """
-check_passed: bool = False
-check True
-    timeout time_add(now(), Fw.TimeIntervalValue(1, 0))
-    persist Fw.TimeIntervalValue(0, 0)
-    freq Fw.TimeIntervalValue(0, 100000):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_multiline_different_order(fprime_test_api):
-    """Test multi-line check with clauses in different order (freq, persist, timeout)."""
-    seq = """
-check_passed: bool = False
-check True
-    freq Fw.TimeIntervalValue(0, 100000)
-    persist Fw.TimeIntervalValue(0, 0)
-    timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_multiline_persist_timeout_order(fprime_test_api):
-    """Test multi-line check with persist before timeout."""
-    seq = """
-check_passed: bool = False
-check True
-    persist Fw.TimeIntervalValue(0, 0)
-    timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_singleline_any_order_persist_first(fprime_test_api):
-    """Test single-line check with persist before timeout."""
-    seq = """
-check_passed: bool = False
-check True persist Fw.TimeIntervalValue(0, 0) timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_singleline_any_order_freq_first(fprime_test_api):
-    """Test single-line check with freq before other clauses."""
-    seq = """
-check_passed: bool = False
-check True freq Fw.TimeIntervalValue(0, 100000) persist Fw.TimeIntervalValue(0, 0) timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_singleline_only_persist(fprime_test_api):
-    """Test single-line check with only persist clause."""
-    seq = """
-check_passed: bool = False
-check True persist Fw.TimeIntervalValue(0, 0):
-    check_passed = True
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_singleline_only_freq(fprime_test_api):
-    """Test single-line check with only freq clause."""
-    seq = """
-check_passed: bool = False
-check True freq Fw.TimeIntervalValue(0, 100000):
-    check_passed = True
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_multiline_timeout_occurs(fprime_test_api):
-    """Test multi-line check that times out."""
-    seq = """
-timed_out: bool = False
-check False
-    timeout time_add(now(), Fw.TimeIntervalValue(0, 100000))
-    freq Fw.TimeIntervalValue(0, 10000):
-    assert False, 1
-timeout:
-    timed_out = True
-assert timed_out
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_mixed_inline_and_multiline(fprime_test_api):
-    """Test check with some clauses inline and some on indented lines."""
-    seq = """
-check_passed: bool = False
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0))
-    persist Fw.TimeIntervalValue(0, 0)
-    freq Fw.TimeIntervalValue(0, 100000):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_mixed_two_inline_one_multiline(fprime_test_api):
-    """Test check with two clauses inline and one on indented line."""
-    seq = """
-check_passed: bool = False
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0)
-    freq Fw.TimeIntervalValue(0, 100000):
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ==================== Check Statement Control Flow Tests ====================
-
-
-def test_check_break_not_allowed(fprime_test_api):
-    """Test that break inside check body is not allowed when check is not in a loop."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    break
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_continue_not_allowed(fprime_test_api):
-    """Test that continue inside check body is not allowed when check is not in a loop."""
-    seq = """
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    continue
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_break_allowed_in_loop(fprime_test_api):
-    """Test that break inside check body IS allowed when check is inside a loop."""
-    seq = """
-loop_ran: bool = False
-while True:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        loop_ran = True
-        break
-assert loop_ran
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_continue_allowed_in_loop(fprime_test_api):
-    """Test that continue inside check body IS allowed when check is inside a loop."""
-    seq = """
-iterations: I64 = 0
-while iterations < 3:
-    iterations = iterations + 1
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        continue
-assert iterations == 3
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_return_in_function(fprime_test_api):
-    """Test that return inside check body works correctly in a function."""
-    seq = """
-def check_and_return() -> I64:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        return 42
-    return 0
-
-result: I64 = check_and_return()
-assert result == 42
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_return_from_timeout_body(fprime_test_api):
-    """Test that return from timeout body works correctly."""
-    seq = """
-def check_timeout_return() -> I64:
-    check False timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) freq Fw.TimeIntervalValue(0, 10000):
-        return 1
-    timeout:
-        return 42
-    return 0
-
-result: I64 = check_timeout_return()
-assert result == 42
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_both_branches_return(fprime_test_api):
-    """Test that a function with check where both bodies return doesn't need trailing return."""
-    seq = """
-def check_returns() -> I64:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        return 42
-    timeout:
-        return 0
-
-result: I64 = check_returns()
-assert result == 42
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_no_timeout_body_return_required(fprime_test_api):
-    """Test that a function with check but no timeout body still needs trailing return."""
-    seq = """
-def check_needs_return() -> I64:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        return 42
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_no_timeout_clause_still_needs_return(fprime_test_api):
-    """Test that a check with no timeout clause still needs a trailing return
-    because the desugared if/else has an implicit else branch that doesn't return."""
-    seq = """
-def check_no_timeout() -> I64:
-    check True:
-        return 42
-"""
-    # The check desugars to: while True:...; if result: <body> else: <implicit>
-    # The implicit else doesn't return, so we need a trailing return
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_check_nested(fprime_test_api):
-    """Test nested check statements."""
-    seq = """
-outer_passed: bool = False
-inner_passed: bool = False
-
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    outer_passed = True
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        inner_passed = True
-
-assert outer_passed
-assert inner_passed
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_nested_timeout_body(fprime_test_api):
-    """Test nested check with inner check timing out."""
-    seq = """
-outer_passed: bool = False
-inner_timed_out: bool = False
-
-check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-    outer_passed = True
-    check False timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) freq Fw.TimeIntervalValue(0, 10000):
-        pass
-    timeout:
-        inner_timed_out = True
-
-assert outer_passed
-assert inner_timed_out
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_inside_while_loop(fprime_test_api):
-    """Test check inside a while loop."""
-    seq = """
-iterations: I64 = 0
-checks_passed: I64 = 0
-
-while iterations < 3:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        checks_passed = checks_passed + 1
-    iterations = iterations + 1
-
-assert iterations == 3
-assert checks_passed == 3
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_inside_for_loop(fprime_test_api):
-    """Test check inside a for loop."""
-    seq = """
-checks_passed: I64 = 0
-
-for i in 0..3:
-    check True timeout time_add(now(), Fw.TimeIntervalValue(1, 0)):
-        checks_passed = checks_passed + 1
-
-assert checks_passed == 3
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_check_with_zero_clauses(fprime_test_api):
-    """Test check with absolutely no clauses (all defaults)."""
-    seq = """
-counter: I64 = 0
-
-def eventually_true() -> bool:
-    counter = counter + 1
-    return counter >= 2
-
-passed: bool = False
-check eventually_true():
-    passed = True
-
-assert passed
-assert counter >= 2
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_func_modify_param(fprime_test_api):
-    seq = """
-def test(arg: U8):
-    arg = 1
-    assert arg == 1
-
-val: U8 = 123
-test(val)
-assert val == 123
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ==================== Simulated Time Tests ====================
-
-
-@pytest.mark.skipif("config.getoption('--use-gds')", reason="simulated time is only available in the Python model")
-class TestSimulatedTime:
-    """Tests for simulated time functionality.
-
-    These tests verify that the sequencer model properly:
-    - Tracks simulated time
-    - Advances time when sleep() is called
-    - Returns the configured timeBase from now()
-    - Correctly handles timeBase incompatibility
-
-    These tests are skipped when --use-gds is passed because they rely on the
-    Python model's simulated time, which cannot be configured on a live GDS.
-    """
-
-    def test_now_returns_initial_time(self, fprime_test_api):
-        """Test that now() returns the configured initial time."""
-        seq = """
-t: Fw.Time = now()
-# Initial time of 5 seconds = 5,000,000 microseconds
-# timeBase=TimeBase.TB_NONE, timeContext=0
-assert t.timeBase == TimeBase.TB_NONE
-assert t.timeContext == 0
-assert t.seconds == 5
-assert t.useconds == 0
-"""
-        assert_run_success(fprime_test_api, seq, initial_time_us=5_000_000)
-
-    def test_now_returns_configured_time_base(self, fprime_test_api):
-        """Test that now() returns the configured timeBase."""
-        seq = """
-t: Fw.Time = now()
-# Configured timeBase=TimeBase.TB_WORKSTATION_TIME
-assert t.timeBase == TimeBase.TB_WORKSTATION_TIME
-assert t.timeContext == 0
-"""
-        assert_run_success(fprime_test_api, seq, time_base=2)
-
-    def test_now_returns_configured_time_context(self, fprime_test_api):
-        """Test that now() returns the configured timeContext."""
-        seq = """
-t: Fw.Time = now()
-# Configured timeContext=4
-assert t.timeContext == 42
-"""
-        assert_run_success(fprime_test_api, seq, time_context=42)
-
-    def test_sleep_advances_simulated_time(self, fprime_test_api):
-        """Test that sleep() advances simulated time correctly."""
-        seq = """
-# Get time before sleep
-t_before: Fw.Time = now()
-
-# Sleep for 2 seconds and 500000 microseconds (2.5 seconds total)
-sleep(2, 500000)
-
-# Get time after sleep
-t_after: Fw.Time = now()
-
-# Calculate the elapsed time
-elapsed: Fw.TimeIntervalValue = time_sub(t_after, t_before)
-
-# Should have slept for exactly 2.5 seconds
-assert elapsed.seconds == 2
-assert elapsed.useconds == 500000
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_sleep_multiple_times_accumulates(self, fprime_test_api):
-        """Test that multiple sleep() calls accumulate time correctly."""
-        seq = """
-t_start: Fw.Time = now()
-
-# Sleep 1 second
-sleep(1, 0)
-# Sleep 0.5 seconds
-sleep(0, 500000)
-# Sleep 0.25 seconds
-sleep(0, 250000)
-
-t_end: Fw.Time = now()
-elapsed: Fw.TimeIntervalValue = time_sub(t_end, t_start)
-
-# Total: 1.75 seconds
-assert elapsed.seconds == 1
-assert elapsed.useconds == 750000
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_time_cmp_same_time_base_works(self, fprime_test_api):
-        """Test that time_cmp works when both times have the same timeBase."""
-        seq = """
-t1: Fw.Time = now()
-sleep(1, 0)
-t2: Fw.Time = now()
-
-# t2 should be greater than t1
-result: Fw.TimeComparison = time_cmp(t1, t2)
-assert result == Fw.TimeComparison.LT  # t1 < t2
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_check_with_different_time_base_crashes(self, fprime_test_api):
-        """Test that check crashes when now() and timeout have different time_bases.
-
-        This tests the full check statement integration: the check desugars to use
-        time_cmp(now(), timeout), and if the time_bases differ, the assert should crash.
-        """
-        seq = """
-# Construct a timeout with a different timeBase than what now() returns
-# now() returns timeBase=TimeBase.TB_NONE by default
-# Set timeout with timeBase=TimeBase.TB_PROC_TIME
-bad_timeout: Fw.Time = Fw.Time(TimeBase.TB_PROC_TIME, 0, 100, 0)
-
-check True timeout bad_timeout persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 100000):
-    pass
-timeout:
-    pass
-"""
-        # Now run with default timeBase=0, but the timeout uses timeBase=1
-        assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
-
-    def test_check_with_simulated_time_timeout(self, fprime_test_api):
-        """Test that check properly times out based on simulated time advancement.
-
-        This test verifies the full check loop:
-        1. now() returns simulated time
-        2. sleep() advances simulated time
-        3. Check properly detects timeout when simulated time exceeds deadline
-        """
-        seq = """
-timed_out: bool = False
-
-# Set timeout to be 100ms from now
-# With freq of 10ms, we'll check ~10 times before timeout
-check False timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    # This shouldn't run because condition is always false
-    assert False, 1
-timeout:
-    timed_out = True
-
-assert timed_out
-"""
-        # Start at time 0, each sleep(0, 10000) advances 10ms
-        # After ~10 iterations, we hit 100ms and timeout
-        assert_run_success(fprime_test_api, seq)
-
-    def test_check_condition_persists_over_simulated_time(self, fprime_test_api):
-        """Test that persist duration is measured using simulated time.
-
-        The check condition must remain true for the full persist duration
-        (measured in simulated time).
-        """
-        seq = """
-# Track how many times the condition is checked
-check_count: I64 = 0
-
-def condition() -> bool:
-    check_count = check_count + 1
-    return True  # Always true
-
-# Require condition to persist for 50ms with 10ms frequency
-# Should need ~5 checks to persist
-check condition() timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 50000) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-timeout:
-    assert False, 1
-
-# With simulated time, we should have checked at least 5 times
-# (initial + enough to accumulate 50ms of persistence)
-assert check_count >= 5
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_sleep_float_advances_time(self, fprime_test_api):
-        """Test that sleep with float argument advances simulated time."""
-        seq = """
-t_before: Fw.Time = now()
-
-# Sleep for 1.5 seconds (1 second + 500000 microseconds)
-sleep(1, 500000)
-
-t_after: Fw.Time = now()
-elapsed: Fw.TimeIntervalValue = time_sub(t_after, t_before)
-
-# Should have slept for 1.5 seconds
-assert elapsed.seconds == 1
-assert elapsed.useconds == 500000
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_now_time_base_preserved_through_check(self, fprime_test_api):
-        """Test that now() consistently returns the configured timeBase throughout check."""
-        seq = """
-# Verify timeBase is consistent inside check
-check_count: I64 = 0
-time_base_ok: bool = True
-
-def check_time_base() -> bool:
-    check_count = check_count + 1
-    t: Fw.Time = now()
-    # Should always have timeBase=TimeBase.TB_SC_TIME
-    if t.timeBase != TimeBase.TB_SC_TIME:
-        time_base_ok = False
-    return check_count >= 3
-
-check check_time_base() timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
-    pass
-timeout:
-    assert False, 1
-
-assert time_base_ok
-assert check_count >= 3
-"""
-        assert_run_success(fprime_test_api, seq, time_base=3)
-
-
-# ==================== Time Function Tests ====================
-
-
-def test_time_function_basic(fprime_test_api):
-    """time() parses ISO 8601 strings to Fw.Time."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00Z")
-# Unix timestamp for 2000-01-01T00:00:00Z is 946684800
-assert t.seconds == 946684800
-assert t.useconds == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_with_microseconds(fprime_test_api):
-    """time() parses ISO 8601 strings with microseconds."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00.123456Z")
-assert t.seconds == 946684800
-assert t.useconds == 123456
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_sleep_until(fprime_test_api):
-    """time() can be passed directly to sleep_until()."""
-    seq = """
-sleep_until(time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_WORKSTATION_TIME))
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_invalid_format(fprime_test_api):
-    """Invalid time string format should fail at compile time."""
-    seq = """
-t: Fw.Time = time("not a valid time")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_time_function_invalid_format_2(fprime_test_api):
-    """Time string without Z suffix should fail."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_time_function_default_time_base(fprime_test_api):
-    """time() defaults to timeBase=0 and timeContext=0."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00Z")
-assert t.timeBase == TimeBase.TB_NONE
-assert t.timeContext == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_custom_time_base(fprime_test_api):
-    """time() accepts custom timeBase parameter."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_WORKSTATION_TIME)
-assert t.timeBase == TimeBase.TB_WORKSTATION_TIME
-assert t.timeContext == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_custom_time_context(fprime_test_api):
-    """time() accepts custom timeContext parameter."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00Z", timeContext=5)
-assert t.timeBase == TimeBase.TB_NONE
-assert t.timeContext == 5
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_all_params(fprime_test_api):
-    """time() accepts all parameters."""
-    seq = """
-t: Fw.Time = time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_SC_TIME, timeContext=7)
-assert t.timeBase == TimeBase.TB_SC_TIME
-assert t.timeContext == 7
-assert t.seconds == 946684800
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_named_args(fprime_test_api):
-    """time() works with named arguments."""
-    seq = """
-t: Fw.Time = time(timestamp="2000-01-01T00:00:00Z", timeBase=TimeBase.TB_PROC_TIME)
-assert t.timeBase == TimeBase.TB_PROC_TIME
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_time_function_negative_seconds(fprime_test_api):
-    """Time before Unix epoch (1970) should fail with negative seconds error."""
-    seq = """
-t: Fw.Time = time("1969-01-01T00:00:00Z")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_time_function_u32_overflow(fprime_test_api):
-    """Time after year 2106 overflows U32 seconds."""
-    # U32 max is 4,294,967,295 seconds after 1970 = year ~2106
-    seq = """
-t: Fw.Time = time("2200-01-01T00:00:00Z")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# Tests for non-constant-sized types (types containing strings)
-
-
-def test_non_const_sized_var_decl(fprime_test_api):
-    """Variable declarations with non-constant-sized types should fail."""
-    seq = """
-val: Ref.DpDemo.StringArray = Ref.DpDemo.StringArray("a", "b")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_non_const_sized_func_param(fprime_test_api):
-    """Function parameters with non-constant-sized types should fail."""
-    seq = """
-def foo(x: Ref.DpDemo.StringArray):
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_non_const_sized_func_return(fprime_test_api):
-    """Function return types with non-constant-sized types should fail."""
-    seq = """
-def foo() -> Ref.DpDemo.StringArray:
-    pass
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_non_const_sized_ctor_call(fprime_test_api):
-    """Calling constructors for non-constant-sized types should fail."""
-    seq = """
-Ref.DpDemo.StringArray("a", "b")
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_return_nothing_expr_in_void_func(fprime_test_api):
-    seq = """
-def test():
-    return Fw
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_use_loop_var_in_func_before_declared(fprime_test_api):
-    # loop_var is scoped to the for body; functions can't access it
-    seq = """
-def fun():
-    assert loop_var == 2
-
-for loop_var in 0..2:
-    pass
-
-fun()
-"""
-
-    assert_compile_failure(fprime_test_api, seq)
-
-
-# ── Flag tests ──────────────────────────────────────────────────────────────
-
-
-def test_set_flag_basic(fprime_test_api):
-    """set_flag with a FlagId enum constant and True value should succeed."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_set_flag_false(fprime_test_api):
-    """set_flag with False value should succeed."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, False)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_set_and_get_flag(fprime_test_api):
-    """set_flag followed by get_flag should return the set value."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-assert get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL) == True
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_set_flag_toggle(fprime_test_api):
-    """Setting a flag to True then False should work."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-assert get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL) == True
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, False)
-assert get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL) == False
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_set_flag_dynamic_value(fprime_test_api):
-    """set_flag should accept a runtime bool expression for value."""
-    seq = """
-x: bool = True
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, x)
-assert get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL) == True
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_get_flag_in_expression(fprime_test_api):
-    """get_flag result should be usable in boolean expressions."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-x: bool = get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL) and True
-assert x == True
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_set_flag_wrong_type(fprime_test_api):
-    """set_flag with an integer instead of FlagId should fail compilation."""
-    seq = """
-set_flag(0, True)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_flag_wrong_type(fprime_test_api):
-    """get_flag with an integer instead of FlagId should fail compilation."""
-    seq = """
-get_flag(0)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_set_flag_non_const_index(fprime_test_api):
-    """set_flag with a non-constant flag index should fail compilation."""
-    seq = """
-x: Svc.Fpy.FlagId = Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL
-set_flag(x, True)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_flag_non_const_index(fprime_test_api):
-    """get_flag with a non-constant flag index should fail compilation."""
-    seq = """
-x: Svc.Fpy.FlagId = Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL
-get_flag(x)
-"""
-    assert_compile_failure(fprime_test_api, seq)
-
-
-def test_get_flag_assign_to_var(fprime_test_api):
-    """get_flag result can be assigned to a bool variable."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, False)
-flag_val: bool = get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL)
-assert flag_val == False
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-flag_val = get_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL)
-assert flag_val == True
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ── EXIT_ON_CMD_FAIL behavior tests ────────────────────────────────────
-
-
-def test_exit_on_cmd_fail_flag_causes_exit(fprime_test_api):
-    """When EXIT_ON_CMD_FAIL is set and a command fails, the sequence should exit with error."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-Ref.cmdSeq.RUN("", Svc.FpySequencer.BlockState.NO_BLOCK)
-"""
-    assert_run_failure(
-        fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR,
-    )
-
-
-def test_no_exit_on_cmd_fail_flag_allows_failure(fprime_test_api):
-    """When EXIT_ON_CMD_FAIL is explicitly off, a failing command should not halt the sequence."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, False)
-resp: Fw.CmdResponse = Ref.cmdSeq.RUN("test", Svc.FpySequencer.BlockState.NO_BLOCK)
-assert resp == Fw.CmdResponse.EXECUTION_ERROR
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_exit_on_cmd_fail_with_successful_cmd(fprime_test_api):
-    """When EXIT_ON_CMD_FAIL is set but the command succeeds, no exit should occur."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-CdhCore.cmdDisp.CMD_NO_OP()
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_exit_on_cmd_fail_toggle_off_before_cmd(fprime_test_api):
-    """Setting EXIT_ON_CMD_FAIL then unsetting it before a failing cmd should not exit."""
-    seq = """
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, True)
-set_flag(Svc.Fpy.FlagId.EXIT_ON_CMD_FAIL, False)
-Ref.cmdSeq.RUN("", Svc.FpySequencer.BlockState.NO_BLOCK)
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ---------------------------------------------------------------------------
-# Type constructor default values
-# ---------------------------------------------------------------------------
-def test_struct_ctor_all_defaults(fprime_test_api):
-    """Struct constructor with no args should use all defaults from dictionary."""
-    seq = """
-pair: Ref.SignalPair = Ref.SignalPair()
-assert pair.time == 0.0
-assert pair.value == 0.0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_struct_ctor_partial_defaults(fprime_test_api):
-    """Struct constructor with some args should use defaults for the rest."""
-    seq = """
-pair: Ref.SignalPair = Ref.SignalPair(time=1.0)
-assert pair.time == 1.0
-assert pair.value == 0.0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_struct_ctor_override_all_defaults(fprime_test_api):
-    """Struct constructor with all args should ignore defaults."""
-    seq = """
-pair: Ref.SignalPair = Ref.SignalPair(3.0, 4.0)
-assert pair.time == 3.0
-assert pair.value == 4.0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_array_ctor_all_defaults(fprime_test_api):
-    """Array constructor with no args should use all defaults from dictionary."""
-    seq = """
-depths: Svc.ComQueueDepth = Svc.ComQueueDepth()
-assert depths[0] == 0
-assert depths[1] == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_array_ctor_partial_defaults(fprime_test_api):
-    """Array constructor with some args should use defaults for the rest."""
-    seq = """
-depths: Svc.ComQueueDepth = Svc.ComQueueDepth(e0=42)
-assert depths[0] == 42
-assert depths[1] == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_struct_ctor_enum_member_default(fprime_test_api):
-    """Struct with an enum member should be constructable with defaults."""
-    seq = """
-stat: Ref.PacketStat = Ref.PacketStat()
-assert stat.BuffRecv == 0
-assert stat.BuffErr == 0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-def test_array_elem_non_first_struct_member(fprime_test_api):
-    """Accessing a non-first struct member on an array element must not crash.
-    """
-    seq = """
-val: Ref.SignalPairSet = Ref.SignalPairSet( \
-    Ref.SignalPair(1.0, 2.0), \
-    Ref.SignalPair(3.0, 4.0), \
-    Ref.SignalPair(5.0, 6.0), \
-    Ref.SignalPair(7.0, 8.0))
-assert val[0].value == 2.0
-assert val[1].value == 4.0
-"""
-    assert_run_success(fprime_test_api, seq)
-
-
-# ── Multi-line and trailing comma tests ─────────────────────────────────
-
-class TestMultilineAndTrailingComma:
-    """Expressions inside brackets/braces/parens can span multiple lines,
-    and trailing commas are allowed in struct/array/function-call/parameter lists."""
-
-    def test_multiline_anon_struct(self, fprime_test_api):
-        """Anon struct split over multiple lines."""
-        seq = """
-val: Fw.TimeIntervalValue = {
-    seconds: 10,
-    useconds: 500
-}
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_anon_struct_trailing_comma(self, fprime_test_api):
-        """Anon struct split over multiple lines with trailing comma."""
-        seq = """
-val: Fw.TimeIntervalValue = {
-    seconds: 10,
-    useconds: 500,
-}
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_trailing_comma_anon_struct_single_line(self, fprime_test_api):
-        """Trailing comma on a single-line anon struct."""
-        seq = """
-val: Fw.TimeIntervalValue = {seconds: 10, useconds: 500,}
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_anon_array(self, fprime_test_api):
-        """Anon array split over multiple lines."""
-        seq = """
-x: U32 = [
-    10,
-    20,
-    30
-][1]
-assert x == 20
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_anon_array_trailing_comma(self, fprime_test_api):
-        """Anon array split over multiple lines with trailing comma."""
-        seq = """
-x: U32 = [
-    10,
-    20,
-    30,
-][1]
-assert x == 20
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_trailing_comma_anon_array_single_line(self, fprime_test_api):
-        """Trailing comma on a single-line anon array."""
-        seq = """
-x: U32 = [10, 20, 30,][1]
-assert x == 20
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_func_call(self, fprime_test_api):
-        """Function call arguments split over multiple lines."""
-        seq = """
-val: Fw.TimeIntervalValue = Fw.TimeIntervalValue(
-    10,
-    500
-)
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_func_call_trailing_comma(self, fprime_test_api):
-        """Function call arguments split over multiple lines with trailing comma."""
-        seq = """
-val: Fw.TimeIntervalValue = Fw.TimeIntervalValue(
-    10,
-    500,
-)
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_parenthesized_expr(self, fprime_test_api):
-        """Parenthesized expression split over multiple lines."""
-        seq = """
-x: U32 = (
-    10
-    + 20
-)
-assert x == 30
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_nested_braces_and_parens(self, fprime_test_api):
-        """Nested multi-line expressions with braces inside parens."""
-        seq = """
-check_passed: bool = False
-check True timeout time_add(
-    now(),
-    {
-        seconds: 1,
-        useconds: 0,
-    },
-) persist {seconds: 0, useconds: 0} freq {seconds: 0, useconds: 100000}:
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_check_clauses_with_anon_struct(self, fprime_test_api):
-        """Check statement with multi-line anon struct in timeout position."""
-        seq = """
-check_passed: bool = False
-check True timeout now() + {
-    seconds: 1,
-    useconds: 0,
-} persist {
-    seconds: 0,
-    useconds: 0,
-} freq {
-    seconds: 0,
-    useconds: 100000,
-}:
-    check_passed = True
-timeout:
-    assert False, 1
-assert check_passed
-"""
-        assert_run_success(fprime_test_api, seq)
-
-
-# ── Python-inspired multiline / continuation / trailing-comma tests ─────
-# Test cases adapted from CPython's Lib/test/test_grammar.py to verify
-# that fpy's implicit line continuation and trailing commas behave like
-# Python's.  Only tests that exercise patterns NOT already covered in
-# TestMultilineAndTrailingComma above.
-
-class TestPythonLikeContinuation:
-    """Tests inspired by CPython test_grammar.py – backslash continuation,
-    implicit continuation inside brackets/parens/braces, trailing commas,
-    comments inside multiline expressions, and edge cases."""
-
-    # ── Backslash continuation (CPython test_backslash) ──────────────
-
-    def test_backslash_continuation(self, fprime_test_api):
-        """Backslash at end of line continues to the next (CPython test_backslash)."""
-        seq = """\
-x: U32 = 1 \\
-+ 1
-assert x == 2
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_backslash_continuation_in_assignment(self, fprime_test_api):
-        """Backslash continuation across an assignment expression."""
-        seq = """\
-x: U32 = \\
-    42
-assert x == 42
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_backslash_continuation_multiple_lines(self, fprime_test_api):
-        """Multiple successive backslash continuations."""
-        seq = """\
-x: U32 = 1 \\
-    + 2 \\
-    + 3
-assert x == 6
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Trailing commas in function definitions
-
-    def test_trailing_comma_one_param(self, fprime_test_api):
-        """def f(a,): pass  — Python allows trailing comma in single param."""
-        seq = """\
-def f(
-    a: U32,
-) -> U32:
-    return a
-assert f(1) == 1
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_trailing_comma_two_params(self, fprime_test_api):
-        """def f(a, b,): pass  — trailing comma with two params."""
-        seq = """\
-def f(
-    a: U64,
-    b: U64,
-) -> U64:
-    return a + b
-assert f(1, 2) == 3
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Trailing commas in function calls
-
-    def test_trailing_comma_call_one_arg(self, fprime_test_api):
-        """f(1,) — Python allows trailing comma in single-arg call."""
-        seq = """\
-def f(a: U32) -> U32:
-    return a
-assert f(1,) == 1
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_trailing_comma_call_two_args(self, fprime_test_api):
-        """f(1, 2,) — trailing comma with two args."""
-        seq = """\
-def f(a: U64, b: U64) -> U64:
-    return a + b
-assert f(1, 2,) == 3
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_trailing_comma_call_many_args(self, fprime_test_api):
-        """f(1, 2, 3,) — trailing comma with three args (cf. CPython v0/v1/v2)."""
-        seq = """\
-def f(a: U64, b: U64, c: U64) -> U64:
-    return a + b + c
-assert f(1, 2, 3,) == 6
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Multi-term arithmetic inside parens (CPython test_additive_ops) ─
-
-    def test_paren_continuation_complex(self, fprime_test_api):
-        """Deeply nested arithmetic inside parens across lines."""
-        seq = """\
-x: U32 = (
-    1
-    + 2
-    + 3
-    + 4
-)
-assert x == 10
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_paren_continuation_with_operators(self, fprime_test_api):
-        """Mixed operators inside parens across lines."""
-        seq = """\
-x: U32 = (
-    2 * 3
-    + 4
-)
-assert x == 10
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Comments inside continued expressions (CPython test_suite) ────
-
-    def test_comment_inside_paren(self, fprime_test_api):
-        """Comments inside parenthesized continuation (like Python)."""
-        seq = """\
-x: U32 = (
-    # first term
-    1
-    # second term
-    + 2
-)
-assert x == 3
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_comment_inside_braces(self, fprime_test_api):
-        """Comments inside struct literal."""
-        seq = """\
-val: Fw.TimeIntervalValue = {
-    # the seconds field
-    seconds: 10,
-    # the useconds field
-    useconds: 500,
-}
-assert val.seconds == 10
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_comment_inside_func_call(self, fprime_test_api):
-        """Comments inside a multiline function call."""
-        seq = """\
-val: Fw.TimeIntervalValue = Fw.TimeIntervalValue(
-    # seconds
-    10,
-    # useconds
-    500,
-)
-assert val.seconds == 10
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Nested continuation (CPython test_with_statement pattern) ─────
-
-    def test_deeply_nested_continuation(self, fprime_test_api):
-        """Three levels of nesting: parens > call > struct."""
-        seq = """\
-x: U64 = (
-    Fw.TimeIntervalValue(
-        10,
-        500,
-    ).seconds
-    + 1
-)
-assert x == 11
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Empty lines inside continued expressions ──────────────────────
-
-    def test_empty_line_inside_parens(self, fprime_test_api):
-        """Empty lines inside parenthesized expression (Python allows this)."""
-        seq = """\
-x: U32 = (
-
-    1
-
-    + 2
-
-)
-assert x == 3
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_empty_line_inside_braces(self, fprime_test_api):
-        """Empty lines inside struct literal."""
-        seq = """\
-val: Fw.TimeIntervalValue = {
-
-    seconds: 10,
-
-    useconds: 500,
-
-}
-assert val.seconds == 10
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Multiline function definition
-
-    def test_multiline_func_def_params(self, fprime_test_api):
-        """Parameters each on their own line with trailing comma."""
-        seq = """\
-def add(
-    a: U64,
-    b: U64,
-    c: U64,
-) -> U64:
-    return a + b + c
-assert add(1, 2, 3) == 6
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_func_def_and_call(self, fprime_test_api):
-        """Both definition params and call args multiline (CPython common pattern)."""
-        seq = """\
-def add(
-    a: U64,
-    b: U64,
-) -> U64:
-    return a + b
-
-x: U64 = add(
-    10,
-    20,
-)
-assert x == 30
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Named arguments multiline (CPython keyword arg patterns) ──────
-
-    def test_multiline_named_args(self, fprime_test_api):
-        """Named arguments across lines — cf. CPython d11(1, **{'b':2})."""
-        seq = """\
-val: Fw.TimeIntervalValue = Fw.TimeIntervalValue(
-    seconds=10,
-    useconds=500,
-)
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_mixed_positional_and_named_multiline(self, fprime_test_api):
-        """Mix of positional and named args across lines."""
-        seq = """\
-val: Fw.TimeIntervalValue = Fw.TimeIntervalValue(
-    10,
-    useconds=500,
-)
-assert val.seconds == 10
-assert val.useconds == 500
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Continuation with comparison / boolean operators ──────────────
-
-    def test_multiline_comparison_in_parens(self, fprime_test_api):
-        """Comparison across lines inside parens (CPython test_comparison)."""
-        seq = """\
-x: bool = (
-    1
-    == 1
-)
-assert x
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_boolean_in_parens(self, fprime_test_api):
-        """Boolean operators across lines (CPython test_test)."""
-        seq = """\
-x: bool = (
-    True
-    and True
-    or False
-)
-assert x
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    # ── Multiline in control flow expressions ─────────────────────────
-
-    def test_multiline_if_condition(self, fprime_test_api):
-        """Parenthesized multiline condition in if (common Python pattern)."""
-        seq = """\
-x: U32 = 0
-if (
-    True
-    and True
-):
-    x = 1
-assert x == 1
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_while_condition(self, fprime_test_api):
-        """Parenthesized multiline condition in while."""
-        seq = """\
-x: U64 = 0
-while (
-    x
-    < 3
-):
-    x = x + 1
-assert x == 3
-"""
-        assert_run_success(fprime_test_api, seq)
-
-    def test_multiline_assert(self, fprime_test_api):
-        """Multiline expression in assert (parenthesized)."""
-        seq = """\
-assert (
-    1
-    + 1
-    == 2
-)
-"""
-        assert_run_success(fprime_test_api, seq)
