+++ test/fpy/test_time_seqs.py
@@ -12,16 +12,10 @@ import pytest
 
 from fpy.model import DirectiveErrorCode
 from fpy.test_helpers import (
-    assert_run_success,
     assert_compile_failure,
     assert_run_failure,
+    assert_run_success,
 )
-
-
-@pytest.fixture(name="fprime_test_api", scope="module")
-def fprime_test_api_override():
-    """A file-specific override that simply returns None."""
-    return None
 
 
 # ==================== time_cmp Tests ====================
@@ -801,3 +795,453 @@ while current < target:
 assert count == 5
 """
         assert_run_success(fprime_test_api, seq)
+
+# ==================== Time Builtins Tests ====================
+# Tests for sleep, sleep_until, now(), time constructors, and simulated time.
+# Migrated from test_seqs.py.
+
+def test_get_time_member(fprime_test_api):
+    seq = """
+if Fw.Time(TimeBase.TB_NONE, 1, 2, 3).useconds == 3:
+    exit(0)
+exit(1)
+"""
+
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_rel(fprime_test_api):
+    seq = """
+sleep(1, 1000)
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_rel_default_usec(fprime_test_api):
+    seq = """
+sleep(seconds=1)
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_rel_default_sec(fprime_test_api):
+    seq = """
+sleep(useconds=500)
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_rel_no_args(fprime_test_api):
+    seq = """
+sleep()
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_abs(fprime_test_api):
+    seq = """
+sleep_until(Fw.Time(TimeBase.TB_WORKSTATION_TIME, 0, 123, 123))
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_abs_var_arg(fprime_test_api):
+    seq = """
+x: U32 = 123
+sleep_until(Fw.Time(TimeBase.TB_WORKSTATION_TIME, 0, x, 123))
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_abs_var_arg_2(fprime_test_api):
+    seq = """
+x: Fw.Time = Fw.Time(TimeBase.TB_WORKSTATION_TIME, 1, 2, 3)
+sleep_until(x)
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_wait_abs_bad_arg(fprime_test_api):
+    seq = """
+sleep_until(2, 1, 2, 3)
+"""
+    assert_compile_failure(fprime_test_api, seq)
+
+
+def test_time_type_ctor(fprime_test_api):
+    seq = """
+var: Fw.Time = Fw.Time(TimeBase.TB_NONE, 1, 2, 3)
+if var.timeBase == TimeBase.TB_NONE and var.timeContext == 1:# and var.seconds == 2 and var.useconds == 3:
+    exit(0)
+exit(1)
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_aliases(fprime_test_api):
+    """Fw.Time is an alias for Fw.TimeValue, Fw.TimeInterval for Fw.TimeIntervalValue."""
+    seq = """
+# Fw.Time and Fw.TimeValue are interchangeable
+t1: Fw.Time = Fw.TimeValue(TimeBase.TB_NONE, 0, 100, 500000)
+t2: Fw.TimeValue = Fw.Time(TimeBase.TB_NONE, 0, 50, 0)
+assert t1.seconds == 100
+assert t2.seconds == 50
+
+# Fw.TimeInterval and Fw.TimeIntervalValue are interchangeable
+i1: Fw.TimeInterval = Fw.TimeIntervalValue(10, 500000)
+i2: Fw.TimeIntervalValue = Fw.TimeInterval(5, 0)
+assert i1.seconds == 10
+assert i2.seconds == 5
+
+# Cross-alias operations work
+result: Fw.Time = t2 + i1
+assert result.seconds == 60
+diff: Fw.TimeInterval = t1 - t2
+assert diff.seconds == 50
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_get_time(fprime_test_api):
+    seq = """
+time: Fw.Time = now()
+"""
+
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_const_folding_time_eq(fprime_test_api):
+    seq = """
+assert Fw.Time(TimeBase.TB_NONE, 0, 0, 0) == Fw.Time(TimeBase.TB_NONE, 0, 0, 0)
+assert Fw.Time(TimeBase.TB_NONE, 0, 1, 0) != Fw.Time(TimeBase.TB_NONE, 0, 0, 0)
+"""
+
+    assert_run_success(fprime_test_api, seq)
+
+
+@pytest.mark.skipif("config.getoption('--use-gds')", reason="simulated time is only available in the Python model")
+class TestSimulatedTime:
+    """Tests for simulated time functionality.
+
+    These tests verify that the sequencer model properly:
+    - Tracks simulated time
+    - Advances time when sleep() is called
+    - Returns the configured timeBase from now()
+    - Correctly handles timeBase incompatibility
+
+    These tests are skipped when --use-gds is passed because they rely on the
+    Python model's simulated time, which cannot be configured on a live GDS.
+    """
+
+    def test_now_returns_initial_time(self, fprime_test_api):
+        """Test that now() returns the configured initial time."""
+        seq = """
+t: Fw.Time = now()
+# Initial time of 5 seconds = 5,000,000 microseconds
+# timeBase=TimeBase.TB_NONE, timeContext=0
+assert t.timeBase == TimeBase.TB_NONE
+assert t.timeContext == 0
+assert t.seconds == 5
+assert t.useconds == 0
+"""
+        assert_run_success(fprime_test_api, seq, initial_time_us=5_000_000)
+
+    def test_now_returns_configured_time_base(self, fprime_test_api):
+        """Test that now() returns the configured timeBase."""
+        seq = """
+t: Fw.Time = now()
+# Configured timeBase=TimeBase.TB_WORKSTATION_TIME
+assert t.timeBase == TimeBase.TB_WORKSTATION_TIME
+assert t.timeContext == 0
+"""
+        assert_run_success(fprime_test_api, seq, time_base=2)
+
+    def test_now_returns_configured_time_context(self, fprime_test_api):
+        """Test that now() returns the configured timeContext."""
+        seq = """
+t: Fw.Time = now()
+# Configured timeContext=4
+assert t.timeContext == 42
+"""
+        assert_run_success(fprime_test_api, seq, time_context=42)
+
+    def test_sleep_advances_simulated_time(self, fprime_test_api):
+        """Test that sleep() advances simulated time correctly."""
+        seq = """
+# Get time before sleep
+t_before: Fw.Time = now()
+
+# Sleep for 2 seconds and 500000 microseconds (2.5 seconds total)
+sleep(2, 500000)
+
+# Get time after sleep
+t_after: Fw.Time = now()
+
+# Calculate the elapsed time
+elapsed: Fw.TimeIntervalValue = time_sub(t_after, t_before)
+
+# Should have slept for exactly 2.5 seconds
+assert elapsed.seconds == 2
+assert elapsed.useconds == 500000
+"""
+        assert_run_success(fprime_test_api, seq)
+
+    def test_sleep_multiple_times_accumulates(self, fprime_test_api):
+        """Test that multiple sleep() calls accumulate time correctly."""
+        seq = """
+t_start: Fw.Time = now()
+
+# Sleep 1 second
+sleep(1, 0)
+# Sleep 0.5 seconds
+sleep(0, 500000)
+# Sleep 0.25 seconds
+sleep(0, 250000)
+
+t_end: Fw.Time = now()
+elapsed: Fw.TimeIntervalValue = time_sub(t_end, t_start)
+
+# Total: 1.75 seconds
+assert elapsed.seconds == 1
+assert elapsed.useconds == 750000
+"""
+        assert_run_success(fprime_test_api, seq)
+
+    def test_time_cmp_same_time_base_works(self, fprime_test_api):
+        """Test that time_cmp works when both times have the same timeBase."""
+        seq = """
+t1: Fw.Time = now()
+sleep(1, 0)
+t2: Fw.Time = now()
+
+# t2 should be greater than t1
+result: Fw.TimeComparison = time_cmp(t1, t2)
+assert result == Fw.TimeComparison.LT  # t1 < t2
+"""
+        assert_run_success(fprime_test_api, seq)
+
+    def test_check_with_different_time_base_crashes(self, fprime_test_api):
+        """Test that check crashes when now() and timeout have different time_bases.
+
+        This tests the full check statement integration: the check desugars to use
+        time_cmp(now(), timeout), and if the time_bases differ, the assert should crash.
+        """
+        seq = """
+# Construct a timeout with a different timeBase than what now() returns
+# now() returns timeBase=TimeBase.TB_NONE by default
+# Set timeout with timeBase=TimeBase.TB_PROC_TIME
+bad_timeout: Fw.Time = Fw.Time(TimeBase.TB_PROC_TIME, 0, 100, 0)
+
+check True timeout bad_timeout persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 100000):
+    pass
+timeout:
+    pass
+"""
+        # Now run with default timeBase=0, but the timeout uses timeBase=1
+        assert_run_failure(fprime_test_api, seq, DirectiveErrorCode.EXIT_WITH_ERROR)
+
+    def test_check_with_simulated_time_timeout(self, fprime_test_api):
+        """Test that check properly times out based on simulated time advancement.
+
+        This test verifies the full check loop:
+        1. now() returns simulated time
+        2. sleep() advances simulated time
+        3. Check properly detects timeout when simulated time exceeds deadline
+        """
+        seq = """
+timed_out: bool = False
+
+# Set timeout to be 100ms from now
+# With freq of 10ms, we'll check ~10 times before timeout
+check False timeout time_add(now(), Fw.TimeIntervalValue(0, 100000)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
+    # This shouldn't run because condition is always false
+    assert False, 1
+timeout:
+    timed_out = True
+
+assert timed_out
+"""
+        # Start at time 0, each sleep(0, 10000) advances 10ms
+        # After ~10 iterations, we hit 100ms and timeout
+        assert_run_success(fprime_test_api, seq)
+
+    def test_check_condition_persists_over_simulated_time(self, fprime_test_api):
+        """Test that persist duration is measured using simulated time.
+
+        The check condition must remain true for the full persist duration
+        (measured in simulated time).
+        """
+        seq = """
+# Track how many times the condition is checked
+check_count: I64 = 0
+
+def condition() -> bool:
+    check_count = check_count + 1
+    return True  # Always true
+
+# Require condition to persist for 50ms with 10ms frequency
+# Should need ~5 checks to persist
+check condition() timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 50000) freq Fw.TimeIntervalValue(0, 10000):
+    pass
+timeout:
+    assert False, 1
+
+# With simulated time, we should have checked at least 5 times
+# (initial + enough to accumulate 50ms of persistence)
+assert check_count >= 5
+"""
+        assert_run_success(fprime_test_api, seq)
+
+    def test_sleep_float_advances_time(self, fprime_test_api):
+        """Test that sleep with float argument advances simulated time."""
+        seq = """
+t_before: Fw.Time = now()
+
+# Sleep for 1.5 seconds (1 second + 500000 microseconds)
+sleep(1, 500000)
+
+t_after: Fw.Time = now()
+elapsed: Fw.TimeIntervalValue = time_sub(t_after, t_before)
+
+# Should have slept for 1.5 seconds
+assert elapsed.seconds == 1
+assert elapsed.useconds == 500000
+"""
+        assert_run_success(fprime_test_api, seq)
+
+    def test_now_time_base_preserved_through_check(self, fprime_test_api):
+        """Test that now() consistently returns the configured timeBase throughout check."""
+        seq = """
+# Verify timeBase is consistent inside check
+check_count: I64 = 0
+time_base_ok: bool = True
+
+def check_time_base() -> bool:
+    check_count = check_count + 1
+    t: Fw.Time = now()
+    # Should always have timeBase=TimeBase.TB_SC_TIME
+    if t.timeBase != TimeBase.TB_SC_TIME:
+        time_base_ok = False
+    return check_count >= 3
+
+check check_time_base() timeout time_add(now(), Fw.TimeIntervalValue(1, 0)) persist Fw.TimeIntervalValue(0, 0) freq Fw.TimeIntervalValue(0, 10000):
+    pass
+timeout:
+    assert False, 1
+
+assert time_base_ok
+assert check_count >= 3
+"""
+        assert_run_success(fprime_test_api, seq, time_base=3)
+
+
+def test_time_function_basic(fprime_test_api):
+    """time() parses ISO 8601 strings to Fw.Time."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00Z")
+# Unix timestamp for 2000-01-01T00:00:00Z is 946684800
+assert t.seconds == 946684800
+assert t.useconds == 0
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_with_microseconds(fprime_test_api):
+    """time() parses ISO 8601 strings with microseconds."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00.123456Z")
+assert t.seconds == 946684800
+assert t.useconds == 123456
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_sleep_until(fprime_test_api):
+    """time() can be passed directly to sleep_until()."""
+    seq = """
+sleep_until(time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_WORKSTATION_TIME))
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_invalid_format(fprime_test_api):
+    """Invalid time string format should fail at compile time."""
+    seq = """
+t: Fw.Time = time("not a valid time")
+"""
+    assert_compile_failure(fprime_test_api, seq)
+
+
+def test_time_function_invalid_format_2(fprime_test_api):
+    """Time string without Z suffix should fail."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00")
+"""
+    assert_compile_failure(fprime_test_api, seq)
+
+
+def test_time_function_default_time_base(fprime_test_api):
+    """time() defaults to timeBase=0 and timeContext=0."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00Z")
+assert t.timeBase == TimeBase.TB_NONE
+assert t.timeContext == 0
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_custom_time_base(fprime_test_api):
+    """time() accepts custom timeBase parameter."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_WORKSTATION_TIME)
+assert t.timeBase == TimeBase.TB_WORKSTATION_TIME
+assert t.timeContext == 0
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_custom_time_context(fprime_test_api):
+    """time() accepts custom timeContext parameter."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00Z", timeContext=5)
+assert t.timeBase == TimeBase.TB_NONE
+assert t.timeContext == 5
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_all_params(fprime_test_api):
+    """time() accepts all parameters."""
+    seq = """
+t: Fw.Time = time("2000-01-01T00:00:00Z", timeBase=TimeBase.TB_SC_TIME, timeContext=7)
+assert t.timeBase == TimeBase.TB_SC_TIME
+assert t.timeContext == 7
+assert t.seconds == 946684800
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_named_args(fprime_test_api):
+    """time() works with named arguments."""
+    seq = """
+t: Fw.Time = time(timestamp="2000-01-01T00:00:00Z", timeBase=TimeBase.TB_PROC_TIME)
+assert t.timeBase == TimeBase.TB_PROC_TIME
+"""
+    assert_run_success(fprime_test_api, seq)
+
+
+def test_time_function_negative_seconds(fprime_test_api):
+    """Time before Unix epoch (1970) should fail with negative seconds error."""
+    seq = """
+t: Fw.Time = time("1969-01-01T00:00:00Z")
+"""
+    assert_compile_failure(fprime_test_api, seq)
+
+
+def test_time_function_u32_overflow(fprime_test_api):
+    """Time after year 2106 overflows U32 seconds."""
+    # U32 max is 4,294,967,295 seconds after 1970 = year ~2106
+    seq = """
+t: Fw.Time = time("2200-01-01T00:00:00Z")
+"""
+    assert_compile_failure(fprime_test_api, seq)
